Restructure files and remove TomF implementation
It does not seem to be worth it to keep - it's significantly slower than Tipsify
and does not really produce better results. It could be kept as a reference but
it also does not handle degenerate triangles well (which is fixable of course).

It seems better to focus on Tipsify improvements and abandon this one.

Also rename files and functions so that there's just one algorithm for each
procedure.