Add ATVR to vcache analyzer
ACMR is a good metric in general but the optimal ACMR number for a given
mesh depends on the triangle/vertex ratio - ACMR of 0.5 is impossible to
reach for certain meshes. In contrast, if we assume all vertices are
required by a given mesh, ATVR (average transform to vertex ratio) has
to be at least 1.0 (each vertex is transformed once), and represents the
amount of over-transformation due to vertex cache misses.