Implement overdraw analyzer
The way overdraw analyzer works is by figuring out the number of pixels
we need to shade redundantly (or rather the ratio between the number of
pixels we need to shade and the number of unique pixels we cover) when
rendering the mesh from 6 cardinal axes (+-X/Y/Z) using an ortho camera.

We pick camera based on a bounding cube that envelops the mesh (it's a
cube so that different projections contribute to the result failry) and
then render the mesh using a software rasterizer with a depth buffer to
obtain the overdraw counters.

Also change demo program to use overdraw analyzer and to add an
overdraw-only optimization mode where we start with a dummy cluster and
specify that we don't really care about ACMR by setting the threshold to
a high number. This is not very efficient but it demonstrates the
concept.