overdrawoptimizer: Simplify backface handling
Instead of adjusting rasterization logic it's easier to swap v2 and v3;
this does not change depth gradients and means we do not need to adjust
code that deals with fill convention or check half-space equations for a
negative sign.

Also simplify viewport transformations; as a result some views may be
rotated 90 degrees from the "natural" orientation.