overdrawanalyzer: Pre-transform all triangles
A very significant portion of time during overdraw analysis is spent
in the vertex transformation stage; to be more precise, on large meshes
we have a lot of cache misses during each pass.

To solve this problem, we pre-transform all vertices in all triangles
and put the result into a sequentially accessed array - this ensures we
only pay the price of the cache miss once (and even there it's cheaper
because we don't use the cache for both triangle and overdraw buffer).

This reduces rasterization time from 249ms to 179ms on 1M triangle Buddha
model.