pretransformanalyzer: Add pre-transform cache analyzer
Generally vertex fetch goes through a more or less traditional cache.
This cache can have different size, number of ways, replacement policy
and also may or may not be shared with other units, like texture fetch.

Modeling the real cache is subtle and complicated; this provides a
simple approximation that uses a moderately sized direct-mapped cache.
The performance is close to that of a 4-way cache on typical mesh data
and direct mapping is slightly simpler to implement.

The primary metric for the fetch cache is currently 'overfetch' factor -
that is, the ratio between the number of bytes we need to fetch to
transform all the vertices to the optimal number of bytes aka the size
of the vertex buffer. This will depend on the vertex size, cache
alignment and, most importantly, vertex index order.