demo: Add custom .obj parser
This parser only parses geometry and hasn't been ran through address
sanitizer; it handles most errors in the input file except for
out-of-bounds indices.

It's currently a bit slower than tiny_obj_loader but it's 10x smaller,
and we'll optimize it in the later changes as well as making sure it's
robust.