Implement edge collapse simplifier
The simplifier sorts edge collapses by the maximum error and tries to
collapse as many edges as possible to reach the target threshold. The
error is determined based on triangle/vertex quadrics.

It is currently not attribute-aware so we need to weld everything for
the results to be reasonable - this is obviously not a production
solution :)

Currently the "recursive" collapses are not done, which seems to even
out the geometry a bit - this may need to change in the future.