simplify: Add boundary edge preservation
Identify boundary edges as edges that don't have an opposite, and
construct a quadric from the plane going through that edge and
orthogonal to the triangle surface. We assign a large weight to this
quadric to try to preserve edges better.

Topology analysis is currently using std::set which makes it the
bottleneck in simplification; this will need to change.