Implement vertex remapping for VB/IB generation
The issue with existing generateIndexBuffer/generateVertexBuffer was
that it was not easy to adapt it to the problem of reindexing an
existing indexed mesh (as opposed to indexing an unindexed mesh). It
would be ideal to have similar interfaces for doing this but extending
the current approach resulted in redundant functions.

Instead rework this by requiring the user to generate a remap table
(which is actually equivalent to an index buffer for indexing, but
different for reindexing), and then remap VB/IB based on that.

For these functions you can specify indices=NULL to handle unindexed
meshes if you don't want to generate a dummy index buffer just to remap
it.