simplify: Limit the collapse error during each pass
We collapse edges in passes, and after each collapse the vertices become
locked for the duration of this pass. This means that to reach the
collapse goal, we have to look further in the edge list - much further
for aggressive simplification - which results in us picking edges that
are significantly worse than some existing edges that could be collapsed
in the next pass.

To solve this, compute the approximate error goal (this is not quite
precise because of duplicate edges in the list) and limit the error for
each collapse based on this goal.

This provides a tradeoff between quality of simplification and number of
passes (which dictates performance).