index: Add generateOptimalGrid
This function can generate a 2D grid that is optimized for vertex cache
yielding a perfect score (1 ATVR - each vertex is transformed once) by
prefetching each row of a cache-sized block of quads to make sure that
if the cache is FIFO vertices are only evicted from the cache when they
are no longer necessary.

It is not clear whether this kind of geometry is optimal or measurably
more efficient compared to Tipsify-optimized grid in practice; this is
to be determined.