demo: Implement mesh testing for vcache tester
It can now be invoked with a list of .obj files to compute ATVR of
Tipsify and TomF on the real meshes.

This change also switches to Tipsify implementation from the library;
looks like TomF is beating Tipsify on real hardware (perhaps because LRU
matches real hardware better), so it probably will need to be
reintroduced as well in later changes.