vcache: Add linear vertex cache optimization back
This algorithm was originally part of this library, but then it was
removed because on FIFO simulated cache it was consistently worse than
Tipsify, and had comparatively bad performance characteristics.

Based on testing on real hardware, it resembles LRU much more than it
does FIFO, so having this is actually very beneficial, once we solve the
performance issues.

For now just add it back as is; it can be invoked by setting cache_size
to 0 (the algorithm generates sequences that do not depend that much on
the actual hardware cache size; internally we use 16 as it matches the
behavior of AMD/NVidia GPUs reasonably well).