vcache: Fix linear cache optimizer to handle degenerate triangles
The code that updated the LRU cache assumed that all triangle indices
are distinct, which is not always the case. This code probably needs to
be rewritten to be shorter/simpler but for now this fixes the issue,
which allows us to run the stats on all meshes in the dataset.