vcache: Make sure linear cache optimizer is always linear
The biggest weak spot of the linear optimizer performance-wise is the
best triangle search - it scans over the entire triangle sequence
whenever we hit a dead-end.

When we hit a dead-end this means that we have exhausted all triangle
neighbors of the vertices in the cache, so the next triangle is bound to
be a cache miss and in general will start a new disjoint strip.

It's not clear why we would pick a best score triangle for this, and
just cycling through the triangles in the input order guarantees O(kN)
runtime for the algorithm, with k being the cache size.

The effect of this change on cache efficiency is minimal - some meshes
are slightly better, some are slightly worse.

We might explore dead-end stack similar to Tipsify for this later, as it
seems like it might be beneficial to connect to an existing set of
vertices even if they are not in the cache anymore.