vcache: Generate vertex score tables
It seems like the parameters published in the original paper are not
really optimal. To facilitate their tuning, generate them based on the
input configuration so that we can bruteforce the better set of values.

Note that this change also switches to a table-based vertex valence
scoring - this eliminates a potentially expensive sqrt (which is
expensive if it's not inlined). We clamp effective vertex valence to 31
which seems to be enough for all practical uses...