indexgenerator: Fix VB remap when some vertices are unused
The remap table can contain ~0u destination vertices when some of the
vertices from the source stream aren't used (this can happen for indexed
inputs); we used to perform a malformed vertex copy in this case, and
instead just skip the vertex now.

Also unify indexed and unindexed remapping codepaths.