vfetch: Change parameter order to match other functions better
optimizeVertexFetch was the only function where vertex data pointer did
not immediately precede vertex count/size; change this for consistency.

There still is a difference between functions that use vertex_positions
and functions that use vertices in terms of argument order; this is
unfortunate and might need to be fixed later.