demo: Enhance vcachetester for manual cache analysis
Add a new mode to vcache tester where you can specify an index sequence
and the tool will analyse it's cache behavior.

Specifically:

- For each vertex in the input sequence we'll try to see if it has
resulted in a vertex invocation by comparing invocation count on a
prefix of the sequence that has this index as the degenerate triangle.

- For each vertex in the input sequence we'll try to see if it is still
in the cache by adding a degenerate triangle with that index to the
resulting sequence.

This analysis can misbehave in certain cases, since it morphs the input
sequence assuming that the cache behavior is not too surprising, but it
generally works well and can give insight into how the target GPU cache
behaves.