indexcodec: Introduce C++ wrappers for encode/decode
We reuse the existing wrapper structure which means that index buffer
decoding requires a separate allocation/copy when decoding a 16-bit
index buffer.

This is suboptimal but will have to stay as is for now - fixing this
requires changing the API structure to have both 16/32 bit
implementations exposed for C, and it's not clear if we should do it
just yet and if so if we should do it for just index buffer decoding or
not.