demo: Improve vcachetester behavior wrt cache flushes
Some hardware performs a complete vertex cache flush before transforming
a triangle if the triangle doesn't fit into the cache.

In this case for a triangle a b c, it might be that (a a a) doesn't
increase the invocation count (if a is in cache), (a b b) increases it
by 1, and (a b c) increases it by 3 - if b/c aren't in the cache and
cache has room for just one more vertex.

This change adds code to deal with this situation more correctly and
also adds more assertions to validate the logic.