indexcodec: Rotate triangles before encoding
When we encode triangles that don't follow the edge-vertex pattern, we
now try to rotate the triangle to have a more predictable encoding for
the compressor to recognize the patterns better.

The rotation has to preserve the order of "next" vertices to maintain
encoding efficiency so it's somewhat limited; still, this significantly
improves the frequency of vertex 0 encodings (making "next" much more
likely for v0).

This makes the post-zstd size of meshes ~2% smaller.