Remove namespace meshopt from public APIs
The split between meshopt_ and meshopt:: API is more trouble than it's
worth - we need to forward all functions to meshopt:: namespace even if
they have the same interface as their C counterparts, and this creates
confusion in terms of documentation and examples.

To unify the interface this change also removes default arguments -
essentially the only "extras" you get out of C++ interface now is the
templated index types.