indexcodec: Remove versioning for now
It's not clear if we need the version at all and if we do that it needs
to be checked externally.

Specifically, if we want to provide a memory-safe decoder, it will have
to be able to return an error - in this case we'd be able to encode the
version in the stream itself. This simplifies the interface.

Alternatively we could just lock down the encoding format and create new
functions for new formats if necessary; it seems like we don't need to
lock the format down up until 1.0, or at least until the API loses the
experimental status.