indexcodec: Don't encode vfifo=1 for edge-vertex
In an edge-vertex code, it's extremely rare for vertex to be at offset 1
in the FIFO; this means that it's the last vertex we referenced, so it
implies a structure like

a b c
? ? c

Which by itself isn't rare, but in most cases you'd expect c to be part
of an edge LRU - so it's usually redundant.

This change improves resulting size by 0.3%-1% since it allows us to use
one more entry from the vertex FIFO in some cases.