overdrawanalyzer: Switch fill convention to pixel center
Before this change, the rasterizer would rasterize pixels with top left
corners covered by the triangles. This is DX9 filling convention;
DX10/11/GL and all other APIs compute coverage based on pixel center.

Changing this requires adjusting initial half-space equations to use the
center of the first pixel as a baseline, as well as adjusting the
bounds that we are using to fill.

This makes the rasterizer match hardware behavior better.