indexcodec: Optimize encoding for 3-vertex triangles
When we have a 3-vertex triangle, we rotate the triangle to try to make
sure that the first vertex is new (we can always do it if *any* vertex
is new). In case all vertices are old this generally means that the mesh
was optimized for a vertex cache much larger than 16 since we've seen
all 3 vertices before but don't have the edge in the edge buffer - with
correct optimization settings this almost never happens on real meshes.

Because of this, our encoding for the 3-vertex triangle almost always
starts with f0; the aux byte that follows frequently has very biased
distribution - 00 is the most common, followed by 01/10/02/20 in many
cases. There is unfortunately redundancy in the aux byte because vertex
fetch optimization doesn't pre-rotate the triangles so sometimes we
can't replace 10 with 01.

This change merges the aux byte into the second nibble for f0 encoding
if feb/fec are both <4 (so they either encode a new vertex or a very
recent FIFO position). Since it's very rare for both vertices to share a
FIFO position, we have f5, fa and ff values that don't normally come up.

We use fa to encode the case where fea is 0 (new vertex), but feb/fec
don't fit in 2 bits; this means that for these cases the encoding is
close to what we used to have, replacing f0 with fa.

We use ff to encode the case where fea is 15 (not-new vertex). This
rarely comes up but it's important so that we can encode arbitrary
triangles - we can't rely on each 3-vertex triangle having a new vertex.

f5 is reserved and not encoded; there's no obvious candidate for this
value.

This can have anywhere between 0-1.5x savings pre-zstd on real meshes;
the most effective example is that for meshes with disjoint quads, we
used to encode each quad as, roughly, f0 00 20, and now we encode as f0
20, making the result 1.5x shorter. On most real meshes the impact is
1-5%.

Results on the test data set, in bits/index:

buddha.obj: 3.4 -> 3.4
bunny.obj: 3.4 -> 3.4
electric_motor_vintage.obj: 3.1 -> 2.9
nyra.obj: 3.1 -> 3.0
rungholt.obj: 4.0 -> 2.7
sinbad.obj: 3.0 -> 2.9
zeuxcg.obj: 3.2 -> 2.8

In all cases above, there is no regression on pre-zstd or post-zstd
compressed sizes; post zstd gains are pretty minor, up to 3%.