indexcodec: Use table-driven codeaux encoding
The issue with using 2-bit encoding for feb/fec is that some
combinations are incredibly unlikely. Specifically, 1&2/1&3 and the like
are usually encoded as edge-vertex instead of 3-vertex; conversely, we
can use the 4-bit space more efficiently by encoding some cases like 7&8
- these could be encoded with edge-vertex given more edge FIFO bits but
we only have 4 bits so they don't always fit.

This change has a minor gain in terms of compression efficiency -
depending on the mesh it could get up to 1% pre-zstd.