indexcodec: Switch from next-delta to last-delta encoding
It appears that encoding deltas from next, while generally efficient,
runs into issues for some meshes where the index traversal sequence
after vcache and vfetch optimization leaves a large group of triangles
around (passing them along one edge sequence) for a later traversal,
then traverses something else and then comes back. When it does, we get
an index sequence that requires a set of indices that aren't encodable
using extremely short encoding, so we fall back to next-delta, but these
deltas are large!

It's still the case that the indices are close to each other, they are
just far away from next, so encoding deltas between indices results in
more efficient index sequences.

This change thus switches to zigzag-encoded deltas from last explicitly
encoded index. In all tested meshes this is either neutral on both
pre-zstd and post-zstd sizes, or noticeably increases the compressed
sizes by sometimes encoding fewer bytes, and frequently encoding more
predictable bytes (when last-deltas are very small and next-deltas fit
in a byte but use more unique values).

An issue for some meshes is that for sequences that are generally
efficient to encode using next-delta, switching to a distant index and
switching back only requires to encode one large value (the outlier
itself), whereas for last-delta we get two large deltas - one positive
and one negative. This in general penalizes the last-delta but not
enough for it to lose to next-delta.

It's possible to use a hybrid encoding where we use either next-delta or
last-delta based on what's more efficient; this usually results in a win
compared to last-delta but the code gets more complex, so this change
avoids this for now.

This change has the following improvements in terms of bits/triangle:

buddha.obj: 10.24 -> 9.92 (pre-zstd), 6.81 -> 6.33 (post-zstd)
armadillo.obj: 10.68 -> 10.11 (pre-zstd), 4.05 -> 3.01 (post-zstd)
kitten.obj: 10.44 -> 10.09 (pre-zstd), 6.22 -> 5.51 (post-zstd)

In most cases the gains are small, up to a few %, but for large complex
meshes this can save up to 5% pre-zstd and up to 25% post-zstd.