demo: Work around Intel GPU behavior in vcachetester
Some index sequences trigger the assert inva <= invb in the following
circumstance:

For a repeating 1 2 3 index sequence, on a Intel GPU eventually you'll
get to the following state:

inv0=5 inva=6 invb=5 invc=5 inv1=5

In other words, the sequences (1 2 3)+ 1 2 2 and (1 2 3)+ 1 2 3 don't
generate extra invocations for any vertices of the new triangle, but
(1 2 3)+ 1 1 1 somehow does.

It's not clear why this happens but it's easy to work around - if
inv0==inv1 we know we didn't transform any indices - and in fact, this
means we can improve the performance of tester by only evaluating
inva/b/c if we need to.

It's likely that there are still some sequences that can trigger these
asserts but this eliminates the obvious cases.