vcacheanalyzer: Make analysis results closer to GPU behavior
While details of the hardware behavior are not known and vary between
vendors and GPU generations, based on the results of vcachetester it
seems that both AMD and NVidia follow a similar algorithm:

- Triangles are buffered up to a certain number, after which the cache
is flushed
- Within the buffer, a FIFO cache is used to gather vertices to fill a
warp/wavefront
- When a warp is full, it's dispatched for execution; cache is flushed
at this point - new warps don't share data with old warps

There are some pecularities, for example in case of NVidia the 9xx
series has a cache that has 16 vertices but the cache behavior is
somewhat odd - even if a vertex is still in FIFO cache it will not be
reused in some conditions - whereas the 10xx series seems to be able to
reuse any vertex from the current warp, making the effective cache size
32.

AMD seems to consistently follow the outlined procedure with cache size
14, wavefront size 64 and triangle buffer size 128 for multiple GPUs
from GCN era; pre-GCN cards (e.g. AMD Radeon HD 6xxx) seem to follow
a different - possibly more traditional? - algorithm.

Intel more or less has a simple 128-entry FIFO with no obvious warp
boundaries; there is some sort of limitation on reuse but I can't yet
figure out what the pattern is, so for now we'll assume Intel has a
vanilla 128 entry FIFO cache.

This change also removes LRU simulation because GPUs don't seem to
actually use LRU, so it's not very useful.