indexcodec: Put codeaux table in the padding bytes
Since we need padding bytes anyway, might as well put the encoding table
in them. This makes the encoded format stable wrt future changes in
codeaux table and also makes it possible to use different tables for
different meshes.

A downside of this change is that the table isn't very compressible, so
it might lose up to 16 bytes post-zstd (which doesn't seem like a big
deal).

This change does not switch to dynamic codeaux table, it just encodes
the static one - while some meshes see 1-5% gains when computing codeaux
table dynamically, this complicates the encoder since it requires to
either encode the mesh twice (2x encoding time), or to encode it and
then re-pack the codeaux bytes and move the varint bytes to fill the
gaps, which is somewhat cumbersome.

Finally, we only *really* use the first 14 entries of the table, so the
last two entries are zeroes - this can be used in the future if we ever
need to version the format or need these bytes for other reasons.