demo: Work around Intel cache behavior in vcachetester
On degenerate inputs, Intel GPU sometimes has peculiar results that
aren't understood by vcachetester logic.

Specifically, for a triangle like this:

	2 1 2

We try to replay these sequences:

	2 2 2
	2 1 1
	2 1 2

And in some cases first and third sequences cause +1 invocation but the
second sequence doesn't. The specific logic that hardware follows here
is somewhat puzzling but for now we can work around this by detecting
this pattern and assuming that the third vertex got transformed.