vcache: Cleanly split two vcache algorithms into two functions
meshopt_optimizeVertexCache now doesn't take a cache_size argument and
should be preferred in general because it produces more optimal results
on real meshes & real hardware.

meshopt_optimizeVertexCacheFifo runs up to 3x faster but requires
apriori knowledge of the cache size and additionally requires that the
cache follows strict FIFO replacement policy (which is generally not the
case; in particular, on real GPUs warp boundaries penalize this
algorithm compared to the other one).