overdrawoptimizer: Remove cache_size parameter
The difference between overdraw optimization impact with
cache_size=16,24,32 is pretty minimal on real caches (including cases
where cache size is large, like Intel). This is probably mostly
attributed to the fact that hard cluster selection should ideally be
based on disjoint mesh sections, not one-off "triangle didn't fit"
events, which works reasonably well with any decent cache size - and the
soft boundary is a very approximate heuristic, that tries to break up
large clusters but isn't as sensitive to the exact split positions as
you might expect.

Since we've removed the cache size from the preferred cache optimization
algorithm it makes more sense to just use a fixed internal cache size,
as it doesn't matter all that much and it simplifies the interface.