stripifier: Add initial implementation of triangle strip generator
The algorithm assumes that after vertex cache optimization, the mesh is
sufficiently local such that a short lookahead distance is enough to
gather useful strip data. Additionally since reordering is only
performed within a short distance, the vertex cache performance will not
suffer too much.

The algorithm proceeds by greedily selecting the strip triangles,
matching the edge winding; once we're out of choices, it picks the start
of the next strip by finding the triangle with the lowest valence vertex
(this is close to the heuristic from the SGI algorithm from 1996).

When starting a new strip, we rotate the triangle to try to make sure
that the strip can be continued with one of the buffered triangles to
avoid excessive amounts of single-triangle strips.

In practice the algorithm seems to produce strips that are ~60-65% the
size of the baseline triangle list, with ACMR 10-20% worse depending on
the cache configuration.