stripify: Implement strip swapping
A frequent occurence in meshes during stripification is that a specific
strip ends in a triangle [a b c] and we'd like to use the edge [b c] to
continue but there are no triangles like that left, while there are
triangles with [a c] edge.

What we used to do in this case is to just restart the strip, but that
requires 4 new indices; instead, we can rebuild the strip as follows:

	a b c ->
	a b a c ->
	a b a c d

This replaces the triangle [a b c] with a sequence [a b a], [b c a], [a
c d] - [b c a] is a rotation of [a b c] so this is equivalent to the
initial sequence with the added [a c d] triangle, and one degenerate
triangle - as a result we add 2 indices to build the next triangle
instead of 4, and can continue the strip further.

This requires maintaining a 3-index strip history (to avoid reads from
destination that might be slow), and yet another branch that can output
triangles; rewriting previously output triangles can be suboptimal in
terms of memory access so this code might get refactored later.

This significantly increases the efficiency of strips, getting from ~60%
to ~50% in some cases; the resulting ACMR degrades accordingly but this
will let us reduce the buffer size and keep some of the strip length
gains for higher vertex cache efficiency.