stripify: Perform strip swaps without rewrites
Instead of rewriting previously written strips to swap them when the
non-swapped strip order runs into a dead-end, we now predict that we
will run into a dead-end and correct this when writing the original
strip.

This makes the code slightly easier to understand, requires tracking
less state (a future commit will remove the extra strip vertices) and
plays nicely with write-combined memory in case mesh_stripify targets
that.

This change currently increases the number of findStripNext calls since
we do them redundantly, but this can be fixed later.