stripify: Eliminate redundant buffer searches
Since we now use buffer scanning to orient triangles or swap sequences,
we always know what the continuation of the strip should be - because of
this we don't need to search for the next triangle on every iteration.

Additionally, we can remove redundant edge searches internally; for
example, we only need to check for swapped edge if continuation edge
isn't found, and we can early out when checking for next triangle after
a restart if we found one of the edges (this averages ~1.5 calls per
restart).

Together these optimizations result in ~2x speed improvement for large
meshes without compromising quality - the only impact is that in some
rare cases we don't find the continuation strip in the buffer because
the buffer is missing one entry, effectively reducing the buffer from 16
to 15 triangles.