stripify: Change buffer capacity to 8
With strip swapping, the efficiency of stripifer is much better and the
ACMR impact is also more severe because of that; this change switches
buffer capacity to 8 to mitigate this.

As a result, the resulting strips are within 55-60% in terms of index
count compared to triangle lists, and the ACMR penalty is generally 10%
or less. Capacity of 16 produces strips around 55% but can have a more
significant effect on ACMR; 8 seems like a good tradeoff for now.

In the future it would be interesting to consider adaptive capacity
(e.g. keep capacity at 16 but terminate strips early if cache
implications get too severe, or limit the lookahead distance to smth
like 16-strip length).