tools: Add vcachetuner tool
This can be used to find optimal settings for vertex cache optimization
tables using a genetic algorithm. The optimization is pretty efficient
but results depend on whether you want to optimize for all cache
profiles or just one - as submitted the tool optimizes for all vendors
equally to try to get cache-oblivious settings.

Note that it uses OpenMP for performance and requires changing vcache
tables to be thread-local instead of static.