Add meshopt_Buffer<T> class
This is a bare-bones replacement for std::vector that will be used to
gradually replace all uses of that type. It needs to be declared in
public header for now since we don't have a private header all
implementations can use (something to perhaps change in the future, for
now the extra symbols seem trivial).