vcache: Trim vcache optimizer tables
We only use first 16 entries from the cache score table so we only need
16.

For valence, most vertices in real meshes have small valences; less than
1% of vertices in the (large) test mesh set have valence>8. Because of
this it is hard to gather good data about how large valences should
behave so we limit the value to 8.

These changes make it a bit easier to tune the tables.