indexcodec: Make meshopt_encodeIndexBuffer memory-safe
The encoder now does all appropriate bounds checks to make sure that we
either can encode the data safely within the provided space, or return 0
if there isn't sufficient space. This makes it possible to encode data
using more constrained buffers if necessary.

Note that right now providing the exact buffer size is enough to encode
any input since the encoder is single-pass; if we switch to a multi-pass
encoder this invariant might get violated depending on the
implementation.