stripify: Rotate first triangle to minimize next distance
Instead of arbitrarily picking the first edge that has a corresponding
triangle in the buffer, we now analyze all three edges and pick one that
minimizes the triangle index in the buffer.

This minimizes the effect stripification has on ACMR (by limiting the
triangle movement), and additionally improves stripification efficiency
for reasons I don't yet understand.

Typically this change results in ~1% shorter strips with ~1% lower ACMR.