vcachetuner: Switch from GA to simulated annealing
Instead of using genetic evolution, we now use simulated annealing; we
do one mutation at a time, and replace the old element if the mutation
results in a better element, or if it results in a worse element with a
probability that depends on the temperature.

This allows us to get out of local maxima; to keep the evolution, we
simulate several temperature bands simultaneously, and promote the
elements from lower temperatures to higher temperatures. That way if we
discover a new reasonable table, we can keep evolving it in a more local
fashion at higher bands.

We don't need too many temperature bands; to keep the amount of
parallelism high, we simulate multiple mutations simultaneously and
anneal all of them into one result (see `steps` argument).

Annealing seems to progress much faster than GA; let's see if it can
reach similar or better results in the end...