vertexcodec: Use vertex blocks of the maximum size
Before this change, vertex blocks we encode would often be slightly
smaller than 256 vertices because we only had buffer space for 254 in
some cases. Now we make sure that the blocks we generate are 256
vertices in size; the only exception is the last block.

This requires tracking the offset into the prediction buffer that we
need to refill from, and a bit of extra logic to process the results
correctly.

The change is neutral on compression efficiency, but it can help
performance since we can now assume we get blocks with integer group
count except for the last one, which we could round up.