vertexcodec: Implement vertex decoding
The decoder is currently pretty slow and doesn't check for out of bound
access in all cases so it's not memory safe, but it does decode any
encoded content.

To simplify the decoder a bit, we no longer encode vertex block sizes
and except encoder/decoder to use the same logic to determine the sizes;
this may change in the future.

Additionally to make decoding easier the header and group bit encoding
is now aligning content to the left of the last byte to make sure we can
decode bits without worrying about the group sizes.