indexcodec: Make decoder ~10% faster by reducing branch mispredictions
We use a few branches in the decoder loop that are content-based; many
of them are not triggering often and it's more important to keep the
fast path, well, fast.

One exception however was fec==0 branch; in many meshes, this branch
often alternates between taken/not-taken and it's hard to predict this.

Since fec==0 and fec<15 paths are very similar, convert this to
branchless code. Note that we use branchless pushVertexFifo which is a
bit interesting - we *always* save the vertex, which relies on the fact
that if we don't increment the offset, the vertex effectively replaces
the last vertex in the FIFO - which is never referenced since our FIFO
indices always use 15 as a magic encoding for out-of-FIFO data.