vfetch: Implement vertex fetch optimization via a remap table
meshopt_optimizeVertexFetch is very convenient but it doesn't work when
a mesh is split into several vertex streams - for this to work we need
to generate a remap table once, remap the index stream, and then remap
each vertex stream.

To make this use-case possible, add meshopt_optimizeVertexFetchRemap.
This function is currently duplicating the actual remap logic - while
it's possible to share some of that, we can't use meshopt_remap*
functions because that'd introduce a dependency between different TUs.