vertexcodec: Simplify encoding by removing 1-bit groups
Instead of doing this, we now use 0-bit groups - that is, when
bitslog2=0, we encode 16 bytes worth of zeroes.

The number of groups where 1-bit encoding is beneficial is very small -
since 1 is used as a sentinel, it's essentially all zeroes plus 1 byte
for each outlier, but we pay 2 bytes to encode the 16 bits, so it's
often a net loss.

Removing this allows us to also remove the zero-byte vertex block
special case - it was mostly useful for discarding entire vertex
components which doesn't seem *that* important to warrant this encoding;
with a zero 16b group code we can now encode 64b of zeroes in one byte,
and consequently we spend 4b on 256b of zeroes (entire vertex block),
instead of 1b - but we save 1b per each 256b block of data which
generally is a net win, and it slightly simplifies decoding.

This change is generally between -1 and +1 bpv on the resulting data,
and neutral or slightly smaller post-deflate.