vertexcodec: Detect SSSE3 support at runtime via cpuid on MSVC
When MSVC is used to compile the code, we can use SIMD decoder and check
if SSSE3 is supported via a __cpuid bit. This is feasible because MSVC
allows to use SIMD intrinsics without any special compilation flags - we
now detect this and use SIMD implementation with a scalar fallback.

On MSVC, you can also use /arch:AVX or /arch:AVX2 to tell the compiler
to use AVX/AVX2 instructions during normal codegen (including scalar
fp). When this is the case (__AVX__ define), we assume that the code
won't run on pre-SSSE3 CPUs anyway and enable SIMD decoder without a
fallback.

Finally, for GCC/Clang the intrinsic use is coupled with codegen
emitting SIMD code in the optimizer - for example, -mssse3 both enables
SSSE3 intrinsics and gives optimizer free rein to use these
instructions. Because of this, we don't use a scalar fallback for this
case either.