vfetch: Fix overfetch computation for meshes with sparse vertex buffers
Similarly to previous change, meshes that shared vertex buffers with
other meshes had an unreasonably low overfetch stat because it was
computed assuming the entire vertex buffer is used.

Instead, keep track of unique vertices as we process the mesh and use
that count instead of the original vertex count.