demo: Add vertex/index encoding to simplification
Simplify at this point needs to be split into two separate functions -
instead of just simplifying it emulates a complete pipeline with almost
all algorithms this library provides in one... This will be done
separately.

We currently encode IB for each LOD separately, to get a bit more
efficiency but also to clearly show the inefficiency of current index
encoder when combined with current LOD packing. This can be improved in
the future.