tools: Add mesh encoder
It's a simple command-line tool that can be used to compare geometry
encoding efficiency and decoding performance; the tool supports raw zstd
compression, codec or codec+zstd.

Note that the tool doesn't have a file format and doesn't output
compressed results - it's just a show-case, not a production tool right
now.