simplify: Switch to 11-bit radix sort for edge collapse sorting
Sorting is one of the large bottlenecks during simplification; we often
have way more edges than we actually want to collapse. This change
switches to 11-bit radix sort which singificantly reduces the time to
sort collapses without a noticeable impact on the sort quality.

We use 11-bit keys that are taken from high bits of collapse error;
since error is non-negative, we end up using 8 bits of exponent and 3
bits of mantissa which ends up being enough...

On buddha.obj, this change reduces simplification time from 6.5s to 3.6s.