simplify: Fix manifold->seam collapses
When collapsing onto a seam vertex, it's important to keep the
non-remapped index in the collapse structure. Otherwise we can
accidentally remap a manifold vertex to the other side of the seam, or
flip two seam vertices when collapsing seam on seam.

To fix this, we now remap indices immediately before accessing various
arrays like vertex_quadrics or vertex_locked, which makes the logic more
explicit, and also handle the on-seam remap explicitly.