simplify: Use seam edge quadrics to preserve seam geometry better
Topology-driven collapse makes sure that the mesh doesn't get holes, but
doesn't make an effort to preserve seam geometry. This can result in
significant UV distortion in some cases.

Additionally, due to how current collapse constraints work, there's a
possibility that we can collapse a seam loop onto itself in the middle -
specifically, when we collapse an edge between two seam vertices, we
don't make sure that the edge we're collapsing was a seam edge; it
could've been manifold.

This greatly improves if we add an edge-based quadric similarly to
borders; it helps preserve seam geometry, and makes the seam loop
collapses less likely.

We also use a way less aggressive edge weight now since we don't need
this to avoid holes any more.