simplify: Introduce target_error argument to the algorithm
In addition to trying to reach the target index count, we can now also
try to reach the target error. This can be used in several ways:

- Limit the maximum acceptable error to something reasonable and try to
hit the target index count
- Use target_index_count = 0 and simplify as much as possible to the
error bound

The tests set maximum error to 1e-3 for now which is a reasonable
default but still sometimes produces poor simplification; the error
metric will be tuned over time.

The error should be scale-independent since the mesh is rescaled prior
to simplification.