tools/lodviewer: Add visualization of edge loops
Also add draft code for rotating the mesh that isn't hooked up to input
yet.

Note that when we render edge loops we redundantly re-render edges -
this is similar to point kind visualization, and happens because we need
to limit the rendering to vertices that are present in the final output.