simplify: Clean up edge selection and edge loop remap
With the precise selection of seam edges, we can now activate the
condition that eliminates redundant seam edges, which reduces the number
of edges to fill/sort and slightly improves the error heuristic by
making the estimated error limit more accurate.

Additionally clean up code in remapEdgeLoops a bit which seems to make
it a bit faster, and do minor style cleanup.