simplify: Minor cleanup
Remove TODO entries that we don't plan to act upon.

Degenerate triangles in the input aren't an issue since we need to
handle zero-area triangles in a robust manner regardless of whether they
are topologically degenerate.

While Garland97 paper uses Q1+Q2 to estimate the edge error, it doesn't
seem to provide measurable improvements and in some cases actually makes
simplification subjectively worse. In a way, if v1 is already
displaced from Q1, we've already taken this error into account and it's
not clear if we need to continue using this in subsequent collapses.
Additionally, memoryless simplification that we might want to switch to
later tells us that ignoring the original quadrics after each
simplification can produce better results so it's not clear how rigorous
we should be in our attempts to preserve it in this algorithm.