simplify: Use a more precise limit on the number of collapses
Each edge collapse removes 1 or 2+ triangles depending on whether the
edge was a border edge. We now use both the edge collapse and triangle
collapse goals:

- Edge collapse goal is used to approximate a reasonable error bound for
the current pass; this goal isn't precise and is generally conservative

- Triangle collapse goal is used to stop collapsing edges; this goal is
usually precise and is occasionally underestimating the number of
collapses for edges with >2 adjoining triangles

As a result, we frequently see the last pass simplifying down to the
exact target count instead of requiring one or two more tiny passes.