simplify: Make bailout conditions more robust
This change replaces the commit 1ad606b091fd9e312381e13a0b6fd18404894aaa
with a more robust handling of edge cases.

Instead of trying to predict that we've run out of collapses, we try to
perform collapses and bail out if we couldn't get a single one.

This requires less special handling around odd cases like
target_index_count not being divisible by 3 and in general makes sure
that if we need to add collapse rules into performEdgeCollapses, we
don't need to add extra logic to the algorithm itself.