stripify: Add stripifyBound and unstripifyBound
As per a newly established convention, functions that write variable
amounts of data to the user-allocated destination buffer have a Bound
variant that provides worst case output size.

Even though the formulas for worst case size of stripify/unstripify are
simple, it seems worth it to provide these functions for uniformity.