clusterizer: Fix robustness issues by discarding wide cones
If a cone is close to a full hemisphere, then some triangles are almost
perpendicular to the cone axis. This leads to the intersection point
between the triangle plane and the axis being really far away, which can
cause stability issues.

Solve these by discarding cones like this early in the pipeline.
cos(angle) reaches 0.1 at 84 degrees; this leaves 6 degrees from either
side for the negative cone to fit the camera point - which can happen
but is very rare - we shouldn't penalize the culling rates with this
change.