clusterizer: Improve cone efficiency by computing a more precise axis
Instead of just averaging the normals to build the axis, we now use the
bounding sphere center of normals (treated as points on a unit sphere)
to determine the axis direction.

If bounding sphere is computed precisely, then this results in an
optimal cone; however, precise construction is complex and inefficient
so we use an approximate algorithm instead.

The algorithm for sphere construction is described here:
Jack Ritter. An Efficient Bounding Sphere. 1990

This commonly improves the culling efficiency by 20% relative to the
baseline (e.g. 20% rejected clusters -> 24% rejected clusters).