clusterizer: Rename meshopt_Cone et al to meshopt_Bounds
Since for many cases you want to have both backface and bounding sphere
culling, the function will now compute both - this change only changes
the interface, and doesn't compute the sphere yet.

The reason this is helpful is because we actually need to compute the
cluster center anyway so we might as well use the bounding sphere center
for that. This also reduces the API surface area a bit.