clusterizer: Add support for quantized cone normal
In general 8-bit precision should be enough to store the cone normal and
cutoff; we now quantize both to 8-bit SNORM and offset the cutoff to
have a conservative test.

While it would be possible to store the cutoff in 8-bit UNORM format,
this is somewhat less uniform - easy to accidentally use the same
decoding code - and can't be fetched using typed SNORM buffers with
automatic decoding.