clusterizer: Add perspective culling formula that doesn't use apex
Instead of using the cone apex, we can use the cone center and bias the
test by the half of the solid angle of the cluster's bounding sphere.

Ordinarily this would require inverse trig or some substitute to compute
sin(a + b), where a is the cone semiangle and b is the half solid angle.

However, we can use the fact that sin(a + b) <= sin(a) + sin(b), and
conservatively replace the former with the latter. We already have
sin(a) - this is cone cutoff - and sin(b) = radius / distance.

The formula that doesn't use the apex is beneficial if you already are
using the sphere since you don't need to store the 12b apex.