simplify: Simplify position remap building
The position remap code was blindly copied from generateVertexRemap and
adjusted to remap to the first (canonical) vertex with equivalent
position; however, this code is a bit more complex than necessary.

The (obvious in hindsight) insight is that by construction, if index I
is present in the hash set, then remap[I] == I, so we don't need to
query remap - just need to use the hash set key. This simplification has
already been applied to shadow IB generation.