demo: Optimize isMeshValid and areMeshesEqual in Debug
isMeshValid spent a lot of time in operator[] so extract pointers ahead
of time to remove all extra calls. This reduces time from 8.2s to 0.1s.

areMeshesEqual performs deindexing and sorting, both of which are very
slow in Debug. Since we just need to compare meshes, we can compute an
order-independent hash by computing a hash for each triangle and mixing
them together. As a result, the function drops from 68.6s to 2.9s.