tools: Implement material group support in objparser
We now parse usemtl statements in .obj files and create a separate array
of object groups that can be ignored if necessary (we don't need to
change existing code), or can be used by the application.

For simplicity, the length of the material name is limited to 255
characters which should probably be enough.