js: Clean up and rework decoder emcc output
Decoder is now using modularized output with wrapper JS functions that
malloc/free the data. We use emmalloc which is very small, and enable
growable heap to make sure we can process arbitrarily large data.

Closure compiler almost works but it renames decode/encode functions and
I'm not sure yet how to fix this so it's disabled for now.