js: Add OptMeshLoader
This change implements an .optmesh loader for three.js that uses
js/decoder.js to decode vertex/index data with wasm-compiled code.

The interface matches that of OBJLoader; instead of returning each mesh
fragment as a separate Mesh object, we create just one Mesh object with
one VB/IB and use draw groups and interleaved attributes to render parts
of it.

The position data is encoded using normalized int16; to avoid having to
modify shaders, we use a scale for the Mesh and put the Mesh in a Group
so that users are free to modify the scaling parameters of the Group
without disturbing our packing. Ideally we'd inject position and UV
unpacking code into the shaders but that makes it hard to reuse existing
THREE.js materials. This also limits us to 0..1 UVs for now.