vertexcodec: Add NEON support for MSVC ARM targets
We have to work around a few issues to make this happen:

- MSVC doesn't define any GCC-like macros for NEON support, so we have
to use something similar to the SSE detection - namely, MSVC-specific
architecture macros. We assume that NEON is available on both ARM and
ARM64 on MSVC targets which should be accurate.

- MSVC has two separate headers, arm_neon.h and arm64_neon.h, instead of
combining all intrinsics in arm_neon.h.

- MSVC uses a type-agnostic internal vector type, __n128, that is
similar to __m128 for SSE and doesn't support initialization at all. So
instead of initializing byte_mask directly, load it from constant memory
which should result in the same code for GCC/Clang.