vertexcodec: Fix clang-cl build without -mssse3
On MSVC, we use cpuid detection to dynamically select between scalar and
SSSE3 implementations at runtime.

When meshoptimizer is compiled with clang-cl, it activates the same code
path since _MSC_VER is defined, but clang is *not* MSVC so it does not
support using SSSE3 intrinsics without -mssse3 commandline switch, which
generates an error.

It's not clear if it's possible to fix this in clang so for now just
disable the fallback path - to use SIMD decoder with clang-cl you need
to specify -mssse3.