vertexcodec: Optimize neonMoveMask to take advantage of ARM64
On ARM64 we can use vaddv_u8 which computes a horizontal sum of the
entire 8-wide vector; doing this for each half gets us two 8-bit masks
that we need in separate registers anyway. We can only use this on
gcc/clang since MSVC lacks this intrinsic.

On ARM we still use 3 vaddp_u8 to perform two horizontal sums in three
steps, but instead of assembling the result into an endianness-dependent
uint16_t and then decomposing the result back, we just extract two lanes
from uint8x8_t which is simpler and faster.

As a result, both ARM and ARM64 codepaths are ~1-2% faster.