js: Remove targetOffset argument from decoding functions
It's generally expected to be 0, and it's possible to specify a subarray
as a target for decoding, so it doesn't seem worth it to have an extra
output offset (especially as we aren't providing a way to specify the
source offset).

Also rename arguments in the non-minified source to more clearly
annotate the types.