simplify: Introduce a second simplifier, simplifySloppy
The default simplifier faithfully follows the topology and doesn't like
collapsing vertices in ways that violate topological consistency. It is
likely that we'll try to unlock some of the collapse options in the
future somehow, and/or recommend discarding and regenerating some
attributes via shadow IB, but that fundamentally limits the algorithm.

Additionally, the simplifier doesn't know how to collapse small regions
of the mesh other than faithfully following topology - which in some
cases is great but isn't always appropriate. It doesn't merge features
that are topologically disjoint because, again, it's topology-driven.

The new simplifier instead completely disregards topology, using an
out-of-core simplification foundation where the space is divided into
cells, and all vertices in a cell are merged together. We currently pick
the vertex with the minimum quadric error, TBD if solving the quadric
improves anything.

This is an early draft with many shortcomings, in particular the grid
size is currently hardcoded - we'll need to dynamically discover it
based on the target count/error.