simplify: Add more precise triangle counting method and options
For experimentation we now have multiple options for the algorithm:

- Whether we should count target cells or target triangles. Target cells
are a gross approximation, whereas target triangles are always an upper
bound on the resulting mesh complexity.

- Whether we should cache the quadric errors or not. This takes a bit
more memory and is a bit slower for small simplifications since we
immediately compute the error for single-vertex cells, but is a bit
faster for large simplifications since we need to compute half as many
errors.

- Whether we should filter resulting triangles. In some inputs we have
lots of duplicates but usually the duplicate rate is ~5%...