simplify: Implement approximate cell counting
We can get *very* close to the true cell count by using a byte hash
table that's larger than a table that we need strictly speaking (to
reduce collision probability) and counting hits as cells.

We still need to compute the last pass using a real hash table, but
that's just one pass. The downside of course is that there's no way for
us to do correct triangle accounting in this model.