simplify: Implement adaptive merging of cells
We now have an option of creating coarse cell structure on top of the
initial cell grid, where 2x2x2 region of fine cells maps to a coarse
cell. For each coarse cell we compute the number of triangles that we'd
save by merging the cell, and the error based on the quadric data.

The quadric error is approximate since it's based on the fine cell
vertices, not on all vertices in the model.

We proceed to sort the cells in the order of increasing error, and
collapse as many cells as necessary to get us to required triangle
count.