simplify: Remove target_error from simplifySloppy
Also expand documentation for simplify/simplifySloppy and add
ACCELERATE_BINARY option back which allows us to skip binary passes.
It's not clear whether we're going to *really* need binary passes for
performance - I'd like to either figure out a way to only use binary
data (which could be substantially faster), or a way to never use it
(which could be substantially less code).