simplify: Improve interpolation search behavior
We now more carefully spell out loop invariants and clamp the grid size
to min_grid+1 which means that if interpolation search is hesitant to
move forward at all, we will still try to step by 1 grid element - this
can give a slightly better approximation in terms of number of
triangles.

Also limit the number of passes more precisely and exit out of the loop
if we accidentally found the exact number of triangles we needed.