simplify: Add weight to Quadric structure
Each quadric now tracks the weight that it has been initialized with,
and this weight accumulates with quadric addition. This allows us to
compute normalized errors (that reflect squared distance to quadric
plane for simple quadrics) as well as un-normalized errors. Normalized
errors are easier to reason about; it remains to be seen whether we use
them for ranking or just for limiting collapse errors.

We also don't need quadricMul in isolation and now create a weighted
quadric in 4 extra flops instead of 10.