simplify: Implement support for Complex vertex kind
Complex vertices can be collapsed into any target by collapsing all
wedges into a single target one. This obviously ignores attributes but
can preserve the shape better.

After the collapse all future collapses will only need to handle the
wedges of the target vertex in the original topology - so we don't need
to update the wedge lists during simplification.

While Complex vertices could in theory be collapsed arbitrarily,
currently we rely on kind[] information during the collapse, so we only
collapse A->B if B is a weaker kind than B, so for now only allow
Complex->Complex and Complex->Locked collapses.

For now we don't classify any vertices as Complex - that'll have to
depend on simplification options.