js: Reduce stack size to 32K
We actually comfortably fit in 16K but just in case set it to 32K. This
makes sure the code is compatible with new LLVM wasm backend, since it
needs MEMORY_SIZE to be equal to STACK_SIZE plus a small constant.

WASM=1 fixes a warning in emcc; we were always building with it implied.