js: Work around Edge ChakraCore bug in index decoder
WebAssembly version of writeTriangle is miscompiled by Edge
(ChakraCore); the compiler seems to completely drop the `else` branch
from the native code based on the behavior.

Changing `else` to `else if` fixes this issue and doesn't seem to have
any measurable performance penalty; just in case this is only included
for Emscripten builds.