gltfpack: Output enough information to decompress data
The output doesn't follow proper GLTF extension encoding yet but this
makes it trivial to hack GLTF loaders to support meshopt format.

This will need to change if we start using index data for pgram
prediction, which will mean that we'll have to decode on a per-mesh
level, not on a per-buffer level.