gltfpack: Add .obj file parsing
The goal for gltfpack is to be able to produce efficient .gltf/.glb
assets starting from a .gltf input, but it's very convenient to be able
to test .obj files as well.

Additionally, if we can support .obj files, we can completely remove
meshencoder/optmesh in favor of that, and maintain the ability to work
with .obj inputs.

Our .obj parser doesn't support material libraries, so this change adds
fast_obj.h that does (and also is slightly faster than our parser), and
adds support for .obj meshes to gltfpack.

To avoid having to handle gltf data conditionally, we synthesize a fake
cgltf_data structure from .obj file as well.