gltfpack: Add an option for unit-length normals/tangents
By default we output non-unit-length vectors as it improves precision
and reduces entropy - the idea is that conformant renderers need to
normalize the input normal/tangent after scaling anyhow.

However, in some cases renderers might not want to spend the extra
instructions on normalization in which case the newly added option can
be valuable.

Also apparently we didn't *actually* support proper tangent output
before since we accidentally used the same path for normals and tangent
so tangent.w was always 0. This change fixes that as well.