gltfpack: Fix rounding of animation frame count
Instead of ceilf(), let's round to nearest but favor keeping the
animation data over throwing it away.

ceilf() is problematic in that if we very slightly overshoot the integer
frame target due to either a large sample rate difference or a precision
issue - say, the length is 1.0001s at 10 Hz - we'll include an extra
frame which will result in visible issues in looped animations.

We now round the resulting length but limit the amount of animation data
we throw away to 20% of a frame.