gltfpack: Implement animated mesh node analysis
We now recursively tag all nodes with animated ancestors as animated,
and have sufficient scaffolding to keep animated meshes attached to
their original nodes (although this specific part isn't implemented
yet).

To merge meshes, we now are able to analyze the nodes they are attached
to. This means that even if we don't transform all meshes to world
space, we can still merge a fair number of them in common cases (same
parent, no transforms).

Finally we no longer store names for animated nodes because on further
reflection it seems redundant.