gltfpack: Refactor mesh merging
Mesh merging now makes sure that meshes with no indices/streams aren't
left in the target array - instead of leaving unmergeable meshes alone,
we copy them to the first free slot.

This is a bit more expensive - although it could be free with C++11 in
the future - but it simplifies the processing logic elsewhere.