simplify: Implement experimental point cloud simplifier
We use the same quadric based formulation, except that the error
function is the squared distance to the point, and reuse most of the
infrastructure/approach - interpolated+binary search hybrid with the
initial guess.

To count the number of unique points we can't rely on triangle filtering
(which is a bit faster) and have to use the hash table based
deduplication.