gltfpack: Implement fallback scaffolding
When -cf setting is specified, we will save uncompressed binary data to
the fallback buffer so that loaders that don't support the compression
extension can load the data.

This requires adjusting the JSON format for the buffer views; this
change simply implements scaffolding necessary to support this - we now
save an additional .fallback.bin file when requested and reference it
from the JSON blob as a buffer #1.