js: Update to Emscripten 1.37.48
This change switches to the latest implementation of sbrk. The version
we were using before was growing the heap too aggressively which is the
motivation for this change; along with that we're inheriting the
emscripten_resize_heap callback which is just as well since we can now
update the JS heap variable instead of creating a new view every time.

This makes the .js.gz file ~2% larger but hopefully represents a stable
ABI that won't break with future Emscripten versions.

This was built with https://github.com/emscripten-core/emscripten/pull/9636
which is necessary to keep the size increase minimal; building without
this change substantially increases the JS/wasm size and requires a lot
of redundant imports.