vcachetuner: Implement multi-dimensional tuning
Instead of just tuning for vertex cache efficiency with the same weight,
we can now tune for a combination of profiles where a profile can be
tuning for vertex cache efficiency of a specific GPU or for compression
efficiency.

Each profile can also be weighted to provide the optimal balance.