vertexcodec: Improve wasmMoveMask codegen
Instead of computing two 8-bit masks separately, we do this in parallel
using packed 64-bit arithmetics. In theory it's possible to drop down to
32/16-bit with this approach but it's made more difficult by the fact
that this requires a shuffle that isn't representable in SSE.

With this, we can remove the extra branches on any_true as - at least on
Chrome - they aren't providing a performance benefit. As a result,
decodeBytesGroupSimd becomes small enough to be inlineable which gives
us ~10-15% faster decode at the expense of an extra size increase of
~1KB.