gltfpack: Flip the meaning of -vu to unnormalized
By default we now output normalized normal/tangent vectors. As an
option, we still provide -vu which now means "unnormalized" - this
produces smaller compressed files but these files produce validation
issues.

The eventual goal here is to implement unit filters so that we can be
*more* efficient than unnormalized storage and still keep normalized
content. For now, let's default to being correct and have an option of
producing smaller files.