gltfpack: Quantize floating-point values for keyframe tracks
When we can't use fixed point quantization (translation/rotation), we
fall back to floating-point quantization which leaves the least
significant mantissa byte as 0. This results in noticeable improvement
in compression ratio - on two examples, translation track got ~1.6x
smaller (64->40 bits in one case and 40->24 in another).

It remains to be seen how acceptable the quality reduction here is - I'm
worried about translations in particular as there are extreme cases
(e.g. an object moving from 0 to 1000) where this type of change makes
the destination of the movement noticeably off (precision around 1000
drops from 6.1e-5 to 1.6e-2).