gltfpack: Unify embedding flows completely
Now we always load the images into memory when embedding, even when the
source came from a file - so for Basis encoding we load the image into a
file, save it into a temporary file and perform the encoding. This is
somewhat redundant but it cleans up the code flow and given the Basis
encoding performance the extra temporary file roundtrip isn't an issue.

We now only ever call encodeBasisFile when directly encoding data from
file references to file references; this flow will probably get
refactored to go through encodeBasisData as well for consistency...