gltfpack: Fix Basis detection
The code submitted before didn't really work :-/ it's not feasible to
distinguish the absence of Basis executable from the failing run of the
executable reliably. Unfortunately basisu lacks any sort of command that
could be used to run it without it doing anything useful.

For now let's run the executable via popen and confirm that the output
starts with Basis Universal. Eventually hopefully we will be able to use
-version command or similar.