gltfpack: Allow meshes to become empty after processing
We can't output empty meshes so we need to filter them out; several
processing stages could make a mesh empty, so we now split this out into
a separate function.

Additionally, this allows us to separate mesh merging and empty mesh
filtering, which makes merging a bit easier to follow.