vertexcodec: Automatically enable SIMD with fallback for clang/gcc
Starting from gcc 4.9 (advertized in 4.8 but seems to only work from
4.9) and clang 3.8, we can specify a function attribute, "target", for a
specific function to target a specific instruction set.

This means that we can implement SIMD support with scalar fallback the
same way we have been doing that for MSVC - using runtime dispatch, we
can select the right variant to call.

Note that gcc/clang don't have a __cpuid intrinsic. They provide
__builtin_cpu_supports function, however it didn't work correctly in
shared objects in gcc5. This was fixed in gcc6
(https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65612) but it's not clear
which clang version that corresponds to and how to check it...

For now we provide __cpuid implementation using gcc/clang inline
assembly.