indexcodec: Implement last-1 encoding for strips
When using optimizeVertexCacheStrip, we often get long strips that use
existing vertices (think of a sequential strip that goes left to right
in one row, and then right to left in the next row).

For this to be encoded efficiently, it helps to special case the
combination where edge has been encoded previously (strips), and the
vertex is last-1 instead of next.

We steal one value from the vertex FIFO queue which turns out to not be
very important for compression ratio either way, and repurpose this as a
"last-1" code.

Unfortunately doing this requires gating the encode/decode on the
version in a few places. For decode to run as fast as it used to, the
most optimal method seems to be to move the check into the slower path
where we deal with odd vertices. Trying to incorporate this into the
fast path resulted in worse code.