gltfpack: Don't include redundant normal/texcoord streams from .obj
When parsing .obj files, we don't have a specific indication that meshes
include normal/UV data. In fact, it's possible for some faces in the
mesh to have normals and others to not have normals.

This change implements simple filtering that only saves normal/texcoord
streams if at least one vertex in the mesh has the data specified.

This doesn't have a large impact on size when compression is used, but
it's more correct.