gltfpack: Use a separate texture coordinate transform per material
This substantially improves the texture distortion for models that have
different textures that are used with different tiling factors.

This doesn't go as far as it could in terms of improving the quality -
we could split vertices in highly tiled meshes and/or rearrange UV
islands to get a better result, and we could start duplicating materials
to improve this. However, it's a good first step.

Note that this change intentionally keeps a single transform per
material for now to make it easier for renderers to support gltfpack
compression - some renderers support more than one texture matrix per
mesh but not all might.