indexcodec: Improve restart handling a bit
Instead of just resetting fb/fc, we reset the entire vertex FIFO. This
guards against a possibility that one of the vertices after next is
going to refer to a FIFO entry.

While this doesn't result in an invalid stream, this breaks the next
chaining - if we forget to increment next even once, the rest of the
stream will use suboptimal encoding.