js: Switch to -msimd128
This change switches to using -msimd128 compile flags instead of
-munimplmemented-simd128.

The reason for the switch is that in general, compiling with
-munimplemented-simd128 is asking for trouble, as it may generate
instructions that Chrome doesn't understand yet.

We've been carefully avoiding the issues so far, but at this point
-msimd128 supports all but one (swizzle) instruction that we need.

This change explicitly marks the function chain in vertex codec as using
unimplemented-simd128 (you need to mark the chain to enable inlining),
which generates more or less the same code.

The benefit is that codecbench-simd.js can now be ran in Chrome/v8!

Also switch to emscripten_get_now to get more precise timer than
clock().