gltfpack: Implement minification for asset.extras
cgltf doesn't expose the full JSON tree for extras data and instead just
gives us the character range. We currently append it as is to the output
for asset.extras, but that results in odd formatting - we get formatted
JSON in the middle of minified output.

This change fixes that by implementing appendJson that performs
minification of the input stream without parsing it - it's sufficient to
track whether we're inside a quoted string and reject whitespace based
on that.