vertexfilter: Remove redundant WASM instructions
After a fast float->int conversion for signed numbers, we were adjusting
the result to get the byte representation.

First, for values in a range where conversion works (-2^22..2^22), we
already know the resulting exponent so we can subtract 0x4B40_0000
without masking.

Second, because we only need low 8/16 bits anyway, we can skip that as
well.

LLVM was actually eliding the and + sub for us since it knows we only
need the low 8/16 bits so this doesn't affect performance/codegen.