vertexcodec: Add a way to set encoding version
To improve compression ratio, we might need to change the bitstream which
requires introducing a new bitstream version.

While decoding will automatically work with any encoded version, when
encoding we need to specify whether we want a backwards-compatible
result (default) or the one that results in better compression.

This is using the same mechanism we use for index codec. Note that for
now no new version is being introduced.