vertexfilter: Enable variable-bitrate quat coding
Instead of hardcoding 12 bit quaternion components, we now encode 1.f
into the 4th component, similarly to how normals work.

Since we need the low 2 bits for the component index, we only encode the
high byte; this allows us to use 9-16 bits for quaternion components
which is more than enough - 9 bits is typically too little anyhow.

Note that this change makes previously encoded files with -cc
incompatible, but the encoding is much more future proof now.

The quaternion SIMD decoder is ~5% slower but it's stil 2+ GB/s and it
can probably be made a bit faster still.