indexcodec: Implement index sequence encoder
In addition to the triangle index codec, we now have an encoder for a
generic index sequence. The encoder is optimized for small index deltas
and uses *two* baseline indices instead of one - this makes the codec
much more resistant to large deltas particularly for bimodal sequences
where the indices alternate between two more or less sequential streams,
which is surprisingly common.

While the encoder isn't nearly as good as the index buffer encoder for
triangle data, it works pretty well on triangle strip data and really
well on line list data, so it can be used as a fallback.

'D' in the header stands for "delta".