gltfpack: Refactor compression code to support index sequence
Instead of bool compressed we now specify the compression mode as an
enum; this change also reorders the arguments in getBufferView a bit.

We now use index sequence encoder to compress indices for meshes that
aren't triangle meshes; the mesh parsing/processing code still only
supports indexed triangles for now.