gltfpack: Change Mesh to support multiple node attachments
Instead of an optional node pointer we now have an array of nodes.

When merging, the rule is that the node arrays must be pairwise
compatible with each other wrt transforms. This hopefully will maintain
mergeability for common cases without requiring splitting mesh sets.

The rest of the code is adjusted in a trivial manner; when writing node
attachments, we need to write a new quantization node for each target
node because a node can only have one parent.