gltfpack: Rework -kn and -km to explicitly assume naming
The behavior of -kn is still the same as it used to be - it keeps named
nodes and maintains the transform effects of these nodes on the rest of
the scene, including prohibiting mesh merging for attached meshes. The
setting was renamed to keep_nodes to reflect that better.

The behavior of -km is now matching that for materials - it keeps named
materials, prevents merging them with other materials and keeps them in
the file.