gltfpack: Evacuate extras out of JSON blob
Instead of reading the extras data directly out of data->json, we copy
the data we need into a std::string during parsing.

This allows us to reclaim the memory used by the JSON data early,
something that will happen in a separate change.

Ideally we wouldn't need to do this since cgltf could do all of this
itself, but this is a conservative change that doesn't require upstream
work.