gltfpack: Fix a regression with meshes with multiple primitives
This is a regression from ce6dd0c700e0310a8af81db7723f36b0792a9baa. When
parsing mesh primitives, we convert each primitive to a Mesh structure,
and then use node-mesh links to fill nodes[] for each Mesh.

However, we were indexing the mesh primitive array using glTF mesh
index, which is incorrect. We now build a remap table that keeps a range
of Mesh objects for each glTF mesh, and use that to add node mapping to
all meshes in question.