gltfpack: Update README with frequently used options
This change represents a balance between documenting all options in the
README (which feels like excessive duplication) and status quo where
commonly used options may not be discoverable.

The options highlight the most important settings for scene
optimization, and the most important settings that inhibit some types of
optimization that may be needed in certain cases.

Fixes #155.