demo: Update to three.js r118 and start using plugins
Instead of using demo/GLTFLoader.js we now use the plugin infrastructure
added in r118 to load files with EXT_meshopt_compression.

Since three.js is going to deprecate non-module-based examples, the
THREE extension is using `export` as well to conform to the overall style.

Unlike demo/GLTFLoader.js, this method doesn't support mesh instancing
or embedded Basis textures, but this hopefully will be added to
three.js in the future, and for now demo/GLTFLoader.js is still here.