gltfpack: Implement new texture quality scale 1-10
This change changes the texture quality control to use fewer levels,
1-10, with a remap into Basis encoding settings.

The default quality is 8; 9 & 10 are trying to preserve quality more
aggressively, lower levels lose more quality and 1 & 2 are trying to
minimize the size and the encode time as well.

Notably, for UASTC encoding this change enables RDO for all modes; it
does it with a very small RDO dictionary (1024 bytes) because anything
larger than 2048 isn't very practical for encode times.