js: Add meshopt_decoder.module.js
To make it easier to consume meshopt decoder in an ES6 environment with
modules and to make it more compatible with tree shaking, this change
adds a modularized version of meshopt_decoder.

It's the exact same code as the main file, but instead of UMD export it
uses an ES6 export. Unfortunately these can't live in the same file as
export syntax is not valid pre-ES6 and there's no way to specify it
dynamically.