gltfpack: Restore -te option placeholder
A few other options are going to change as part of the upcoming release;
the general strategy is going to be to keep the options in the
deprecated form for a single release, and have them emulate the old
behavior if possible.

This change restores -te and, since there's not a great way to get the
same behavior, just prints an informational message.