gltfpack: Make it possible to specify quantization level
Instead of binary controls over quantization, this change enhances this
to allow three levels of quantization: 0, equivalent to noq; 2,
equivalent to the defaults; 1, where positions and texture coordinates
aren't quantized but normals are.

I'm not 100% sure this is a good choice long term, maybe just using
binary -c and -q options is a better idea, and -noq should be equivalent
to -q 1?