gltfpack: Preserve the time range for pose-only animations
When an animation consists exclusively of constant tracks, we would
previously export only a single keyframe for all tracks. This loses time
information that may be important for playback in case the pose is
sequenced for a given duration by the mixer.

To minimize the size impact, we expand exactly one track and only for
animations that don't have actual timed tracks; for that to work we need
to output a special input track with two time values, start and start +
(frames - 1) / freq, which should match the end time that we would emit
if any tracks had data.