gltfpack: Update required extension checks
We now fail to repack files that already use BasisU textures because
otherwise the texture data would be lost right now; the same is true for
mesh instancing.

If quantization is required then there's no data loss involved, but we'd
expect precision loss and redundant dequantization nodes, so we now emit
a warning.