simplify: Unify two adjacency structures together
This change removes the previously added TriangleAdjacency structure and
replaces it with a change to EdgeAdjacency. Instead of just tracking the
single vertex (edge destination) per edge, we now track the third vertex
of the triangle in the data structure.

This does make the structure larger, and slightly more expensive to
create, but the delta isn't that big - when flips aren't prohibited,
this slows simplification down by ~3%.

However, when flips are prohibited, this makes hasTriangleFlips more
efficient as it doesn't need to use indices or remap; on large meshes
this reduces the number of cache misses substantially, and makes
simplification ~13% faster overall. It's also less code to maintain :)

Note that the meaning of "adjacency" switches after classification in
that we maintain welded adjacency after classification.