simplify: Implement a cheaper version of correct flip detection
In modes 2 & 3, collapses can invalidate the "flip" invariant as long as
they collapse a vertex from the ring of one of the vertices involved in
the original collapse. As such, to prevent collapses it's necessary to
lock the entire ring around each vertex, which is problematic as it
slows down the process even further.

In mode 1, however, because we lock the edge that's being collapsed, all
other collapses can be validated on the fly - as long as we use the
post-collapse vertex position, which we weren't doing before.

This is advantageous as mode 1 has the capacity to be the fastest mode
once we fix the pass scheduling anyway, and so this change removes all
other modes in favor of mode 1 and changes hasTriangleFlips to follow
collapses correctly.