simplify: Add meshopt_simplifyScale for relative <-> absolute conversion
While technically it's possible to compute the mesh extents manually to
switch between relative & absolute coordinates, it's more convenient to
provide a function that computes an error scaling factor.

This can be used to scale input and output errors to meshopt_simplify;
it is also resistant to changes in the scaling logic, e.g. if we ever
decide to use the bounding box diagonal or any other value instead this
will maintain the behavior for "absolute" values.