clusterizer: Implement meshelet score combining distance and cone metrics
buildMeshlets now has an extra parameter, cone_weight, that allows to
balance the importance of cluster size vs cone culling efficiency. When
it's set to 1, size is largerly ignored and we get almost-planar patches
- this is not optimal wrt number of meshlets and can penalize occlusion
culling but works well for cone culling.

When cone_weight is set to 0, cone angle is ignored and we effectively
minimize the cluster size.

When the weight is in the middle, we use a heuristic to balance both
size and cone factors. For this we use a distance normalized using an
expected meshlet size that is computed based on the expected area of
triangles in each resulting meshlet. This is a heuristic but it appears
to work reasonably well in practice and provides a more-or-less smooth
progression in terms of cone culling efficiency (and associated
degradatation of cluster bounds).