clusterizer: Accelerate triangle lookup via a kD-tree
When we run out of connected triangles for a given meshlet, instead of
using a bruteforce search we now use a kd-tree to perform nearest
neighbor lookup in log time.

The code for constructing a tree currently relies on std::sort which
will be fixed in the future. The benefit of a kD-tree here instead of
something simpler like a hash grid is in its ability to handle queries
at various detail levels, although the benefit of this on real scenes
needs to be evaluated in practice.