clusterizer: Change tree layout to a depth-first one
Instead of storing offset divided by 2 in a somewhat awkward fashion we
now store the size of the left subtree, which results in a depth-first
order. One other benefit here is that this allows us to store more data
after leaves if we need to, unlocking leaves that store multiple points.