clusterizer: Optimize statistics gathering by folding divisions
Instead of dividing individual components by the current index, we're
multiplying by the inverse that gets computed at the end of the last
iteration. This replaces 3 divisions by 1 division and 3
multiplications, and importantly improves scheduling since division is
not as much on the critical path.

This makes the entire clusterization process 3-4% faster.