extern: Fix fast_obj path handling for mixed slashes
On Windows, paths with mixed slashes weren't processed correctly as /
would be picked if it was anywhere in the path for the purpose of
determining base.

This also removes the #ifdef _WIN32 from the logic; this helps on
platforms like Emscripten where running the resulting binary that wasn't
compiled with WIN32 still needs backslash processing; paths with
backslashes on Linux should be exceedingly rare, plus we *already*
correct backslashes with forward slashes on Linux.