gltfpack: Fix handling of absolute paths on node.js
When running gltfpack via a node.js build, absolute paths don't work
correctly, as /foo/bar is converted to foo/bar by wasi-libc.

This has to do with the path handling inside wasi-libc, where it looks
at the mounts; given an empty mount path it seems to treat /foo/bar as
mount-relative *and* returns the leading slash which would make sense
for non-empty mount paths.

To work around this, we use the recently introduced chdir support in
wasi-libc. A fake path /gltfpack-cwd is registered as current, and
accesses to that path map to "", which seems to preserve relative paths.

To that end, library.js now implements a slightly more generic version
of mounts, and a stub for path_filestat_get which is needed to get chdir
to work.