clusterizer: Rework meshlet interface to use SoA results
The existing interface for meshlet construction used fixed size arrays
as part of meshopt_Meshlet structure. This was hard to extend - on AMD
hardware larger meshlet sizes may be appropriate, and increasing the
size of the arrays in the structure results in subtantial increase in
memory consumption for older limits.

In addition, meshopt_Meshlet should be regarded as an intermediate
structure as it's not appropriate for direct consumption on GPU, but
that's in contrast with other algorithms in meshopt that usually produce
data fit for GPU.

This change thus reworks the meshlet algorithms to work with packed
vertex index / short index data. This allows us to be flexible wrt the
max meshlet size.

When writing out vertex indices, we pack them tightly; for meshlet
indices we align each meshlet to a 4 byte boundary to maximize GPU
efficiency.

The caller must now allocate 3 arrays. To make the worst case bound
analysis easier, we make sure the array size is simply max_meshlets
times max_vertices / max_triangles*3; to make sure 4b padding doesn't
violate the worst case bound we restrict max_triangles to be divisible
by 4.