indexgenerator: Implement support for triangle adjacency index buffers
Several algorithms that use geometry shaders to render data require
adjacency information; triangle-with-adjacency topology in various APIs
provides a way to supply 3 extra vertices per each triangle that
represent vertices opposite to each triangle's edge.

This data can then be used to compute silhouettes and perform other
types of local geometric processing.

This change implements the data preprocessing step that is similar to
the previous tessellation IB - but instead of storing vertices that
adjoin the triangle from the opposite side, we need to store
complementary / opposite vertices of adjacent triangles, which requires
a separate temporary table.