gltfpack: Adjust animation code to produce deviation
Instead of aborting track scanning on the first delta that's over
tolerance, we just compute the worst case delta. This can be a bit
slower but in practice it's not significant on the profiles, and this
makes other code easier to tweak.

This change also updates the rotation tolerance to exactly 0.1 degrees
(the previous tolerance was a bit smaller).