gltfpack: Compress textures in parallel
Instead of encoding one image at a time and letting Basis do internal
threading we now encode images in parallel using two job pools, one for
actual encoding and another one for farming out Basis encoding jobs.

As a result the compression is much faster, eg jrat.gltf encodes in ~20
seconds before this change and in ~7.5 seconds after it on Ryzen 5900X.

Note that we use two pools but their use isn't very well balanced;
depending on the mode we aren't running as fast as we possibly could, eg
changing both pools to max out logical processor count gets us to 5.5
seconds on the same input. However this would mean that we'd overload
the system so for now we're going to stay conservative.