Update to WASI SDK 14
This switches the builds to latest WASI SDK & LLVM. This required minor
tweaks in WASI stubs to compile.

As a result, the .js files unfortunately get ~10% larger. However, the
vertex decoder is ~10% faster as well, so I'm not very motivated to
chase down the precise source of the difference. The 10% size increase
for decoder translates into 600 extra bytes fetched (gzipped).