Add LICENSE.md to npm packages
Unfortunately npm doesn't seem to allow using external files as part of
the package, so we need to copy the file into the folders before
building, which makes it so that npm picks it up.

This change also moves the build actions that actually produce NPM
artifacts to release workflow to more cleanly isolate that from
build/test.