gltfpack: Switch to normalized UV coordinates
Currently gltfpack quantizes positions and texture coordinates using
integers; this results in very small scaling values that are required to
transform the values back into 0..1 range.

While this is generally not a problem, there are some contexts in which
the scale values are too small and may be truncated to 0, eg if the shader
is using medium precision. Additionally, for native GPU backends like
WebGPU the shader may need to match the logical type of the vertex
attribute with the shader declaration - for unnormalized integers to
work, one must use i32/u32 types, which makes meshes that use
quantization incompatible with shaders that expect non-quantized inputs.

Using normalized data fixes this. For texture coordinates, we should be
able to do that unconditionally. For positions, this can have a downside
in that it requires an application of an extra normalization factor
whenever you work with the data on the CPU side, which can break
frameworks like Three.JS that can read data directly out of Int16Array
otherwise.

This change switches texture coordinates to normalized values, positions
will follow separately.

A downside of this change is a small precision loss during
reconstruction - previously we did the reconstruction in one step, thus
losing ~0.5ulp of precision on that, and now reconstruction is two-step:
GPU divides the integer by 65535, and then we rescale back, which may
result in artifacts - however, in practice the extra 0.5ulp is always
insignificantly small to be observable.

Note that even if input UV is 0..1, with the default settings (12 bits
for quantization) the scale ends up being ~16. Preserving this is
important because it keeps the meshopt encoding efficiency compared to
an alternative where we set least significant bits to 0 instead of most
significant bits (as we do for colors).