simplify: Mark main simplifier as stable
While the simplifier algorithm is still being developed, and extra
functions will be added in the future to work with attributes, it's time
to mark the actual simplify (and simplifyScale) functions as stable.

Sloppy algorithm still may require additional inputs like options in the
future, so keep that experimental for now.