vertexcodec: Use a more efficient lowering for neonMoveMask
Instead of using horizontal adds or more expensive paired adds on 32-bit ARM,
we can use the same trick we use for Wasm which assumes that all input
components are 32-bit and uses a specially crafted constant to turn the result
into a bitmask.

This results in ~2% faster decoding on Apple M2; the performance will need to
be validated on other ARM CPUs.