gltfpack: Implement a test version of parallel exponent encoding
Here instead of sharing the exponent between components of each vertex,
we share the exponent between all vertices for each component, so we
only get 3 unique exponents (and as such the exponent byte compresses
perfectly).

This is similar in spirit to integer quantization, although just like
with all other exponent encodings / float quantization, the position of
the mesh affects the precision of the encoding, not just the scale.