vertexfilter: Add EncodeExpMode to meshopt_encodeFilterExp
Before this change, encodeFilterExp assumed shared exponent encoding.
This is not always the best choice for quality (separate exponent
encoding is preferred there) or compression ratio (for this, it's better
to share exponents across all values for the same input component).

With this change the encoding mode is specified as an input. Some types
of input, like colors or directions, may benefit from SharedVector
whereas other types of input like positions would likely be best
compressed with SharedComponent. As usual, multiple chunks filtered
with different parameters can be concatenated and decoded as one.