indexgenerator: Optimize meshopt_remapVertexBuffer for deinterleaved streams (#562)
When vertex_size is small, the memcpy dispatch overhead is significant
and it dominates the cost of the copy loop. To avoid this, we specialize
the loop for a few common sizes, mostly focusing on deinterleaved
streams as this is where the overhead is highest. More vertex size
specializations can be added in the future if requested.

For gltfpack, this results in ~5% speedup on end-to-end processing of
geometry heavy scenes like Amazon Bistro. meshopt_remapVertexBuffer sees
a ~4x speedup in isolation.