simplifier: Remove compile-time dependency on attribute count
All attribute and quadric data is now supplied independently to relevant
functions. As a side-effect, we're now splitting position and attribute
quadrics entirely - this is technically not necessary here, but it might
be necessary to fix issues with weight normalization and is definitely
going to be necessary to fix issues with error output.