simplify: Reduce the amount of overallocation for collapse array
While some meshes can indeed have 3 collapses per triangle, most meshes
require much fewer. On a typical smooth mesh, every edge occurs twice
and as such we only need half as much memory as we typically use.

This change estimates the capacity that we need for the collapses. It
assumes that every simplification pass can't make the topology require
more collapses than it did initially, and computes an overly
conservative bound by approximating the amount of manifold edges from
the amount of manifold vertices. There are some meshes for which a more
precise calculation (done on triangle-edge level, not vertex level)
yields a smaller estimate, but the vertex-level computation is ~4-5x
faster on average.

As a result, for ~0.1% performance cost on large meshes, we reduce the
total required memory by ~20% assuming a smooth mesh input.