simplify: Free remap position table early
Remap position table is only needed briefly at the beginning of
simplification, but it contributes to the peak memory consumption. This
change introduces a way to deallocate individual allocations (the
deallocation has to be in stack order so that the backing allocator can
simply adjust the bump pointer) and uses it in simplifier and
indexgenerator.

For simplifier, this saves 6-7 bytes per vertex on average, adding up to
~4% smaller memory footprint for large meshes.