gltfpack: Implement support for floating-point texture coordinate quantization
Similarly to position quantization, we can use quantized floating point
values to represent texture coordinates. This can be useful as it trades
off a little bit of memory and compression efficiency for convenience
and interoperability - without the use of texture transforms it becomes
easy to swap textures, reuse materials, and avoid edge cases in
quantization when UV bounds are very large.