gltfpack: Parse individual objects in .obj separately
We now assign a node to each individual object in .obj file; this allows
us to preserve object names. When -kn is not specified, gltfpack will
merge the meshes with the same materials anyway so the result should not
change, but with -kn the nodes will be preserved in the output file.