gltfpack: Merge identical texture objects together
Blender glTF exporter (and possibly others) output duplicate texture objects
in some exports. This blocks material merging which can result in more draw
calls than necessary, and in some loaders (such as Three.JS) may result in
extra texture decoding during loading and extra memory spent.

For simplicity we still assume that images in the source document are
unique, but deduplicate and remove redundant texture objects now.

The process is a little different from how materials are deduplicated, as
we need to maintain remap for unused textures to avoid traversing all views
again as they are part of a complicated material structure.