gltfpack: Merge instanced meshes when instance transforms align
Before we were able to merge meshes when they were parented to the nodes
with the same transform (or rather nodes without transform parented to the
same node), but when instancing was enabled, we'd gather instances and
then refuse to merge the resulting meshes.

Normally this isn't a problem as usually mesh merging works for single
meshes, but this can impact instancing efficiency for structured scenes.