gltfpack: Warn when -kn is used together with -mm/-mi
Since mesh merging and mesh instancing optimizations are disabled when -kn
is used, we now warn when they are specified to avoid confusion. This can
be silenced by omitting -mm/-mi.

Note that in the future it's possible to fix this interaction, as -kn only
promises preservation of named nodes and as such we might be able to still
optimize meshes attached to unnamed nodes, although for -mi to work with
the current implementation the entire hierarchy must be unnamed which is
unlikely.