gltfpack: Change the defaults to copy textures to output
To fix consistency issues with output by default, we now default to
copying (.gltf)/embedding (.glb) textures with or without compression.

There are cases when the previous behavior is desireable. This change
adds a new option, -tr, that keeps the references to original textures
without processing. Notably, this also works for .glb outputs, allowing
to avoid embedding textures into .glb.

When compressing textures, -tr is not currently supported because we do
need to produce new file data. In the future we could consider
supporting -tr by saving the .ktx2 files side by side with the original
images.