gltfpack: Evaluate translation error relative to parent node scale
Before this change we evaluated translation using the node's scale, even
though the node's own scale gets ignored when transforming the node's
translation.

This could lead to issues where gltfpack would ignore the translation
animation in nodes with very small scale even though the parent's scale
was reasonable and the translation was noticeable as a consequence.