simplify: Replace point quadrics with reservoirs
While quadrics are very important for triangle mesh simplification, they
are decidedly overkill for point cloud simplification. We encode a
simpler function into them compared to triangles, and even that function
is too complicated for what we need - minimizing the sum of square
distances to all points in the cell ends up yielding approximately the
same result as averaging all points in the cell and minimizing the
distance to that average.

The benefit of a simpler formulation is that there's less memory used
for encoding the function and less computation required to aggregate the
points / minimize the error.

This also sets us up for including color error into the calculation (by
computing a weighted average of the colors and minimizing the 6D
distance to XYZ+RGB). Even once we include RGB into the reservoirs,
we'll be at 7 floats per cell, vs 11 floats per cell that we spend right
now for weighted quadrics.