simplify: Add color support to simplifyPoints
Following the new reservoir code, we can now easily accumulate colors
into the reservoirs, and add the color error to the position error with
a user-supplied weight. The color error is squared distance; as such,
we scale it by the squared weight.

In addition, we now use color alpha as a point weight; the weighted sum
will thus bias towards more opaque points.

Note that the alpha is not accounted for during point selection.