simplify: Only use RGB colors for simplifyPoints for now
There are a few different options for how we should support alpha during
point simplification, and they are all difficult to evaluate properly
because point clouds with meaningfully diverse alpha that would affect
the simplification noticeably are vanishingly rare.

Additionally, it might be useful to support float3 inputs for densely
packed arrays, as that may make glTF integration easier -- gltfpack pads
all RGB colors with 1.0, but not all other processors might.

Since meshopt_simplifyPoints is experimental we can wait until we get
feedback on this and potentially change it, or maybe introduce
meshopt_simplifyPointsWithWeight in the future where we could supply
weights as a separate channel (and the callers could decide whether to
use alpha or not?).