gltfpack: Use NULL for null pointer constants instead of 0/nullptr
This fixes instances of Wzero-as-null-pointer-constant warning with the
exception of occurrences inside third-party code (extern).

Note that we previous used 0/nullptr/NULL in gltfpack in different
places. gltfpack is C++11 but since it's developed alongside
meshoptimizer which doesn't currently require C++11, for now we use NULL
for consistency instead of nullptr.