js: Expose meshopt_simplifyPoints as meshopt_simplifier @ simplifyPoints
We now support point simplification via a JS interface, both with and
without colors. The order of arguments in the JS wrapper is set so that
simplification without points can just omit the last 3 arguments,
although the resulting order is a little odd when colors *are* supplied.

The function returns the point indices; it's up to the application to
compact the point streams - for triangle meshes we provide compactMesh,
however that only remaps the index buffer. Point vertex data compaction
should be easy enough to implement with a JS for loop.