spatialsort: Mark meshopt_spatialSortRemap as stable
This function has been working well for several years and several
releases, and the interface or implementation is unlikely to change
dramatically.

The implementation could be changed to use a tree based sort in the
future, although experiments in the past suggested that this doesn't
measurably improve the resulting order.

Note that meshopt_spatialSortTriangles is still experimental. It is not
used currently by gltfpack, and it sometimes regresses the compression
ratio; it also might need to start respecting topology in some way.