js: Mark simplifyWithAttributes and simplifyPoints as experimental
Both functions are marked as experimental in C++ code; their interface
and implementation is subject to change. In JS, if the interface of
these functions changes, it may lead to subtle or not-so-subtle issues
with these functions. For now we require the caller to opt into the
use of these functions by setting a module-level experimental boolean.