gltfpack: Update help to account for vtf/vnf and remove vtn
On further reflection, vtn was added prematurely: it only allows setting
the texture coordinate compression mode to default which isn't actually
useful until we add automatic modes, and once we do a lot of things will
change.

This was not very obvious when only vpf and vtf existed, but with vnf as
well it's definitely not a good idea to also add vnn...

Removing this should hopefully be fine since it was a no-op argument that
wasn't documented or advertised anywhere but the command line.