Update documentation to mention max_triangles%4==0
max_triangles technically doesn't have to be aligned to a boundary of 4,
however because meshlet's triangle data is aligned to 4 bytes for GPU
efficiency, if max_triangles is unaligned then it might require
overallocating the output triangle buffer for edge cases, which is error
prone (and complicates the usage requirements of buildMeshletsBound).

This only matters in theory for NVidia that recommends a
slightly-not-power-of-two 126 on the higher end to make sure the extra
metadata (vertex/triangle count) that is implicitly added to the packet
doesn't go over a 128 byte boundary unnecessarily; using 124 in practice
is just as good as 126.

Fixes #643.