Add MESHOPT_SOVERSION CMake option to specify soversion
This patch is taken from Debian patches:
https://salsa.debian.org/roehling/meshoptimizer/-/blob/master/debian/patches/0001-Add-SOVERSION-to-library.patch?ref_type=heads

Specifying soversion currently is at the guidance of the packager;
experimental APIs are not ABI stable but non-experimental APIs are ABI
stable. Since gltfpack uses experimental APIs as well, it probably means
that distributions need to bump soversion on library updates in
practice.

Fixes #644.