clusterizer: Optimize triangle reordering
Instead of tracking visited triangles, maintain an invariant where an
increasing prefix of the triangle array is ordered properly. While it
would be possible to maintain this by swapping the first triangle with
the one we picked at every iteration, this distorts the selection order
in an odd way that may reduce the quality of the generated sequence, and
memmove is fairly fast in practice given that we've already scanned over
the region we are about to move anyhow.

This makes optimization ~40% faster while producing the same results.