js: Cache unpack(wasm_*) result for MeshoptDecoder
This reduces the cost of useWorkers by ~0.3ms on Firefox: unpack isn't
ran often enough to be carefully JIT compiled. This normally doesn't
really matter much but unfortunately by introducing useWorkers API we
gave people the power to call it on every decode, which is a bad idea
but this at least mitigates the cost somewhat.