extern: Improve EXT_mesh_gpu_instancing validation in cgltf
With this change we validate that instanced meshes have a mesh, have
attributes and that all attributes agree on the number of elements.

This makes sure that using the instancing data is less error prone and
is unlikely to lead to memory safety issues. The same change was
submitted upstream: https://github.com/jkuhlmann/cgltf/pull/252