gltfpack: Relax node/instance processing rules
Previously, we had an invariant where a mesh had either nodes but no
instances, instances but no nodes, or nothing (in which case it was
treated as root-attached). We now relax the processing to allow meshes
to have both nodes and instances.

When instances are specified, we still rely on mesh.scene, so instances
are implicitly scoped to a single scene root (and also can't be animated
via their node ancestors).