demo: Implement a basic example of cluster simplification
This is an important building block for Nanite, as it requires a
repeated application of buildMeshlets => simplify (with additional
algorithms that meshoptimizer currently doesn't support, like grouping
meshlets based on proximity). This can be used to refine future
improvements to simplification like sparsity, as well as serve as a
basic example.

For simplicity we use LockBorders flag instead of locking boundary
vertices manually via meshopt_simplifyWithAttributes; from production
perspective both options are viable.