simplify: Add SimplifyErrorAbsolute flag to treat error as absolute
When using sparse simplification, the error is treated as relative to
the mesh subset. This is a performance requirement as computing the full
mesh extents is too expensive when the subset is small relative to the
mesh, but it means that it can be difficult to rely on exact error
metrics.

There are also cases in general, when not using sparse simplification,
when an absolute error is more convenient. These can be achieved right
now via meshopt_simplifyScale but that is an extra step that is not
always necessary.

With this change, when meshopt_SimplifyErrorAbsolute flag is used, we
treat the error limit and the output error as an absolute distance (in
mesh coordinates), and convert it to/from relative using the internal
scale factor.