Minor formatting changes
meshoptimizer.h is not covered by clang-format and it had a couple cases
where spaces were used for indentation instead of tabs.

allocator.cpp *is* covered by clang-format and it's hitting the bug
https://github.com/llvm/llvm-project/issues/34171; it's probably not
worth trying to work around this for a single file that is basically
never edited, but it's valuable to keep `make format` be a no-op.