Allow overriding MESHOPTIMIZER_EXPERIMENTAL externally
Similarly to MESHOPTIMIZER_API, we now define MESHOPTIMIZER_EXPERIMENTAL
only if an existing definition is absent.

This allows, for example, setting this to a deprecated attribute which
still compiles the library itself cleanly, but will warn for any use of
experimental APIs for dependent code. Another interesting use case is
building .so where the experimental symbols are not exported, which
usually results in an ABI stable interface.