vertexcodec: Work around clang analyzer false positive
clang analyzer doesn't realize vertex_count <= vertex_count_aligned
holds (which it does unless addition overflows, which is impossible with
size_t due to a memory limit on vertex inputs, but also we bounds check
vertex count!), and then thinks that we might be writing less than
vertex_count elements to the output buffer. This change fixes this, and
also undefs a forgotten define as a drive-by.