demo: Add a variant of simplifyAttr test with skipped attribute
The construction of inputs to this test actually only requires one of
the attributes to be present, as the middle edge of the mesh has a
gradient shift on any of them, so this doesn't change the simplification
behavior but will help make sure a future optimization preserves the
behavior here.

Also change processDev to call simplifyAttr for mesh testing.