simplify: Only allocate and compute quadrics for used attributes
If some input attributes have weight=0, we allocate and compute quadrics
for those regardless of the fact that we do not need this attribute.
This makes simplification more expensive when used with many attributes
some of which are disabled via weight=0, requiring the caller to
restructure the input data for optimal processing.

With this change, we remap attributes during rescaling and recompute
attribute_count to make this easier. In the future we could also detect
attributes that are constant across the entire mesh, as these also
should not be necessary to take into account.