Add -mcpu=mvp flag to Wasm builds for compatibility
Starting from clang 16, the default Wasm features used silently include
sign-extension which is not part of MVP and is not supported by some old
browsers/devices like iPhone 6. This currently only affects
meshopt_encoder module which is typically not used on device, but a
future C++ or a compiler code change could silently introduce these
instructions in a way that is difficult to test.

For now just disable this since it's not helpful for performance or code
size. We could revisit this in a few years, but clang's policy of
updating features seems irresponsible as it is not actually favoring
compatibility by default.

This change doesn't rebuild js/ binaries, that can happen separately.