Update CONTRIBUTING.md
Add a section to contributing guide regarding the agreement to distribute the code under the terms of MIT license and transfer various rights to the project maintainer.

Prior art (using the same language):

- https://github.com/baldurk/renderdoc/blob/v1.x/docs/CONTRIBUTING.md#copyright--contributor-license-agreement
- https://github.com/ocornut/imgui/blob/master/docs/CONTRIBUTING.md#copyright--contributor-license-agreement

This change naturally does not retroactively affect existing contributions (although GitHub terms of service already establishes this partially); the vast majority of the code is written by the project author anyway. There are no future plans to relicense the project or do anything else that affects the users of the project in any way. This change, however, sets the project up for consistent ownership and allows companies to sponsor development of the project via contracts without the code written as part of that being owned by the sponsor.