simplify: Add minor clarifications to quadricFromAttributes
Slightly expanding on comments for derivation; also note a missing
optimization wrt weight scaling - this is very immaterial for
performance (moving scaling outside of the loop does not make this
faster) or precision, so let's leave it as is.

Note: as part of this validation, in addition to revalidating the
derivation (which can be error prone), this code was also validated by
computing the gradients (gx/gy/gz/gw) using Eigen as a numerical solver.
To do this we can follow the original paper by Hoppe and solve a 4x4
linear system, or solve a 3x3 subproblem that only includes gx/gy/gz
and compute gw so that gx*v0x + vy*v0y + gz*v0w + gw = a0.

This produces more or less the same results during simplification which
proves that the derivation is correct; the analytical formulation this
code uses is much much faster, numerically stable and precise (yielding
~5-10x smaller error as measured by the linear system, as in |Ax-B|).

Since we probably won't need to adjust this in the future, this change
does not include the actual Eigen code used for validation.