simplify: Improve quadricError precision for attribute gradients
Instead of accumulating attribute gradients one sum component at a time,
we can compute the attribute change independently. This allows us to
remove one multiplication, which doesn't really save time in practice,
but more importantly this improves precision of accumulation as it
reduces the chance of cancellation; when tested on color+normal
attributes, we get ~1.5x reduction in floating point error, as measured
by comparing the result of quadricError with the same code using double
precision.