simplify: Switch attribute error to unnormalized area weighted quadrics
Before this change the attribute error used unnormalized edge length
weighted quadrics. The edge length carried over from position error
(where both edge length and area weighting work but edge length at some
point was judged to be slightly higher quality), and 'unnormalized' part
was considered a temporary limitation to be investigated.

With this change, we are officially removing normalization (with good
reason!) and switching to area weighting (with good reason!). In
aggregate this change alters the error and simplification behavior.

The reason why weight normalization is valuable for position quadrics is
it gives the error a particular meaning (a square of the distance
between vertices and surfaces) that is mostly preserved throughout
quadric accumulation. This leads to the intermediate and final collapse
errors being explainable and useful outside of the collapse ordering,
for example the result can be used to select optimal LOD switching
distance.

If applied to attribute quadrics, however, the attribute error becomes
the square of attribute deviation; this value has a limited utility in
isolation, and does not work at all when added to the position error
since the spaces are incompatible. Even if we looked at it in isolation,
measuring position deviation from surface is a measure of silhouette
deformation, whereas a significant attribute error on a very small
triangle is invisible and a small attribute error on a very large
triangle can noticeably affect the appearance.

The correct/ideal way to compute attribute error seems to be to measure
the attribute delta multiplied by the area/distance that this delta is
spread upon. Unfortunately, this does not seem to be representable in a
quadric (further research pending on alternatives) as it's a 4-degree
polynomial if we assume we need squared error * squared distance.

The reason why our previous code worked to some extent is that it scaled
the attribute error by a value that depended on the triangle size; while
this does not correctly work through quadric accumulation, this can
otherwise reasonably approximate the error... except that we used a
linear scale on top of a square of the attribute deviation, and
positional error is internally stored as quadratic, making these
incompatible.

This in turn led to inability to tune attribute weights in a
scale-invariant way or explain what they mean. With large weights, the
combined error was also basically useless for LOD selection, which in
some engines necessitated patching the simplification code to output
just distance based error for LOD selection, which has its own issues.

These issues mostly go away once we accept that unnormalized attribute
metric is a feature and lock the weight to be the triangle area (or
really any other quadratic quantity, eg we could take squared triangle
edge length along the direction of largest attribute deviation, but for
now we're using the area for simplicity).

With these, the attribute weights are easier to explain; a deviation of
1 in a weighted attribute along length L is perceptually equivalent to a
positional shift of length L. If an attribute is 0..1, a weight of 1 is
now reasonable because eg it suggests that a change from white to black
color is the same as just removing the geometry. Of course, smaller or
larger weights still make sense (for example to accentuate the change in
the attribute), and if the attribute is of non-unit scale it should be
counter-scaled via attribute weight as before.

This also fixes the relation between positional and attribute errors and
makes them meaningful to add, as attribute errors are (squared)
distances times attribute deviation. The combined error value is likely
directly usable for LOD selection whereas it was not before depending on
the selected weights.

The one remaning issue is that the metric we use only approximates our
ideal and the approximation breaks down after a series of quadric
accumulations during collapses, as we lose the sense of the aggregate
vertex movement. This might be possible to improve in the future.

This change significantly changes the meaning of attribute weights, but
mostly in a way that should be reasonably easy to adjust for:
applications that used very small (eg 1e-2) weights for unit attributes
should adjust them to be closer to 1; applications that already used
attributes around 1 for unit-length attributes can likely stay as is as
they will get a more balanced position vs attribute handling but can
choose to adjust the weights upwards.