demo: Implement auto-update for simplification module
Instead of having to manually click "update module" every time the
module changes, we now support auto update checkbox that polls the file
every second and reloads it when the modification date has changed.
This, combined with running 'make' in a shell in a loop, results in a
nice hot reload support without too much work.