demo: Expand debug visualization to show all classifiers
Since we have issues with rendering points using our current method, we
just render edges: either when they are marked as loops in the metadata
(which is the case for seam/border), or if both vertices are locked.
When only one vertex is locked, this can sometimes get classified as a
loop (in which case we render it) but in general there is some ambiguity
for edges adjacent to locked vertices, and isolated locked vertices may
not get rendered as such.