simplify: Add a more verbose tracing mode for edge collapses
This is too noisy for any reasonably sized model but it helps debug
small isolated test cases and unit tests. Ideally we would probably
remove the 'edge eval' (as in a given pass we may evaluate way more
edges than we end up collapsing) but for now since we are not storing
position/attribute errors separately that's the only place where we can
output both errors as well as evaluate bidirectional flips.