demo: Adjust simplifySeam test a bit to correct border misconceptions
The seam vertices here have Z coordinate that is higher than zero;
because of this, positional errors are not exactly zero. This is a good
thing since it allows us to get a more stable result across fp
environments and algorithm changes. We adjust the data a little bit to
make the bump less pronounced and fix the comments accordingly.