simplify: Adjust quadricFromTriangleEdge to be slightly more precise
Instead of normalizing p10 we can correct the projection instead: p20p
is effectively scaled by length, and since we multiply it by p10 the
result is scaled by lengthsq. We could divide p20p by lengthsq to
compensate but it's better to instead scale p20 as this avoids
singularity by taking advantage of the normalization that follows.

This should make the function a little faster; it also reduces the
average error by ~1.5x (computed by comparing sqrt(quadric error) for p2
with the double precision distance-to-line computation) on a test model.

comment