gltfpack: Expose simplification error via command-line
Instead of using a hard-coded error, we now allow specifying it via -se
command line argument. This can be useful to allow more significant
simplification when the default 1% error is too restrictive, as well as
relax the limit in combination with -sv where the error could be a
little higher.

Note that by default, we use 1% error for edge simplifier and 10% error
for sloppy simplifier if -sa is specified. This change preserves this
behavior, but when an error is specified we use it for both simplifiers.
So technically -se 0.01 -sa results in a different behavior from just
specifying -sa, but that seems preferable to always amplifying the error
for sloppy simplifier.