gltfpack: Use both normals & colors if present for simplification
Since the core algorithm now automatically reallocates attributes
optimally when some weights are zero, we can fill a temporary array with
a fixed stride of 6 with normals, colors, or both, and pass it along
without having to worry about efficiency. The performance impact of
creating the temporary data is minimal compared to the actual
simplification, and this improves appearance on meshes with color and
normal data further when -sv is used.

For now we use normal weight 0.5 and color weight 1.0 to match demo and
because this works reasonably well, but bumping the normal weight to 1.0
may make sense in the future.