indexgenerator: Add templated overload for meshopt_generateProvokingIndexBuffer
This is useful for signed indices as well as 16-bit indices, although
this comes with a caveat: an input 16-bit index buffer might not fit in
16 bits after the conversion if the provoking buffer size becomes larger than
64K.

This is avoidable if the caller checks the bounds, and 16-bit support
might still be beneficial. For extra safety we guard this with an
assertion.