indexgenerator: Streamline non-provoking remap pass
Instead of manually handling two vertices during a remap, we can iterate
by skipping provoking vertices if we jump in steps 1/2/1/2.

This makes the code a little easier to change in the future if we need
to make the remapping logic more complicated.

Also change the remap logic to match other places in the library where
we update remap conditionally and then write the entry.