demo: Add provoking vertex processing to the demo program
In addition to running the code, we validate that the generated data is
correct (by checking the provoking invariant and checking that the
reorder table recovers the same mesh) and also output the number of
"extra" vertices: given a non-redundant index/vertex buffer, the best
case output from the algorithm is max(vertices, triangles), but
depending on topology it may generate more.

The "extra" percentage may be thus misleading in some cases; for now we
explain that it may not be attainable in a comment, maybe we'll compute
a more precise lower bound in the future.