Revert "demo: Rework .obj reindexing to use fastObjIndex data"
This reverts commit d643b0b44f9ac16f8177e0f01fa65955e486010e.

This code was originally written for gltfpack; while it does indeed run
faster due to a partial reuse of .obj indexing data, it runs into
problems on many real world .obj files where the vertex data is
redundant inside the file, and the resulting index is suboptimal.

This is not an issue for gltfpack because it performs mesh reindexing as
part of mesh processing pipeline, but here that step is absent.