demo: Add nanite.cpp along with -n (make nanite)
This is a stub that will be expanded to provide a demo of Nanite-like
(hierarchical clustered level of detail) processing. This is necessary
both to serve as an example of how to implement it using all best
practices as well as a testing harness; right now meshoptimizer has
enough functionality to implement *a* pipeline, but for best performance
currently some algorithms need to be swapped out. Long term the goal is
for this to be close to optimal while using just meshopt_ functions.