demo: Implement initial hierarchical clustering
This should be a more-or-less complete, if basic, merge-simplify-split
pipeline for cluster DAG build, with the exception of tracking the
actual DAG data (errors and parent links).

The caveat is that to merge clusters, we simply partition them
sequentially; this is suboptimal and where METIS can help. We also use
meshopt_buildMeshlets that can sometimes leave unwanted gaps but we
don't have a great way to quantify the issues yet without collecting
stats.