demo: Implement debug LOD cut output
To make it easier to understand the results, in addition to numerical
stats for the DAG we can now output a simplified mesh by computing the
error for each cluster from a viewpoint and using hierarchical
information to effectively select LODs at every level.

We always do this to output a number, but also can save the .obj to
stderr for subsequent external visualization.