demo: Instead of using a precise merged bounds, merge bounds manually
This falls out of monotonicity requirement: it is not enough to make
LODBounds::error monotonic, the real requirement is that for any
viewpoint, boundsError is monotonic through the DAG. To achieve that we
need to make sure that the bounding sphere of any parent cluster
contains the bounding sphere of the child cluster, which may not hold if
the parent sphere is computed precisely based on vertex data.

Fixing this and fixing boundsError to return FLT_MAX if the viewpoint is
contained inside the sphere makes the DAG checks pass, so they can now
use assertions instead of logs.