demo: Implement initial version of Metis partitioner
We use a k-way partitioning scheme and ask for ~c/4 partitions in hopes
that it groups clusters reasonably well; for now we use the number of
shared edges as the connection weight although it is likely that the
number of shared vertices is a reasonable proxy that is easier to
compute and more useful.

Note that the rest of the pipeline was structured to deal with a more
strict partitioner, as such in some cases the results are better and in
some they are worse. This is good because the rest of the pipeline needs
to have better heuristics anyway.