demo: Add the average tri/cluster counter to nanite
The "full" counter is a little less helpful for METIS which very often
has almost-full clusters, so add an average as well to gauge the
distribution.

We also now count the number of singleton clusters; these are rare but
it's important that they stay that way for the quality of the partition.