gltfpack: Add build support for new basisu library
A few small changes are necessary to compile the newly added code and to
adjust to the new name of the parameter we were setting.

Note that the new version re-adds some debug prints that will need to be
removed, and the threading patch needs to be backported for performance
to be comparable; otherwise it seems to work correctly though.