gltfpack: Use UTF-8 encoding during gltfpack processing
When main() is used, the inputs are ANSI according to the current code
page; this is fine if the file paths can be converted to ANSI, but may
fail to open some files and will use ANSI file paths inside .gltf files
which violates glTF specification.

This can be worked around by using UTF8 as ANSI code page via a
manifest; however, to avoid having to bundle a manifest we can also
convert the command line arguments to UTF8 and set UTF8 as the codepage
instead.

This requires Windows 10 1803; on earlier Windows versions setlocale
should silently fail so this should be more or less equivalent to the
previous code.