simplify: Avoid early termination of passes due to error_goal
In large meshes with complex topology we occasionally hit a case where
error_goal cutoff is not monotonic: a previous pass would simplify up to
a certain goal, but a subsequent pass would arrive at a smaller goal
(for example as invalid collapses get eliminated) and hit it before an
error that it already encountered.

If we already collapsed an edge with a given error it is always correct
to collapse other edges with a smaller error. This change tweaks the
conditional, which results in fewer passes on some meshes (eg
thai_buddha gets ~10% faster to simplify), 2-3% smaller error on some
meshes, and occasionally 1-2% larger error; most meshes simplify as they
did before however.