demo: Adjust visualization to render border-locked edges symmetrically
Border-locked edges as well as seam-locked edges have mismatched kinds
and we currently render them with whatever kind the first vertex has;
this can be misleading as it suggests a different behavior whereas it's
really the same.

For now we will render them using the non-locked kind; currently
simplifier refuses to collapse them but hopefully that will change soon.