simplify: Fix edge loop remap creating invalid loops in some cases
When the seam edge is collapsed in the opposite direction to the loop
flow, we need to follow the previous link to discover the new
connection. However, if that link pointed to a remapped vertex, we
previously would create an invalid loop edge. This would result in
inconsistent loop metadata between loop & loopback, and could prevent
future seam collapses around the problematic area.

The correct thing to do is to remap the connection, unless the loop
stopped after that vertex in which case our vertex becomes terminal.