simplify: Enable seam->locked collapses
Similarly to border->locked collapses, enabling seam->locked collapses
is important to reduce topology based restrictions as that allows us to
reach lower triangle counts or same triangle counts at lower error for
meshes with islands of irreducible topology.

A seam-seam edge is bidirectional whereas a seam-locked edge is
unidirectional and may be connected to any wedge out of an arbitrary
count in the locked vertex. Because of this we can't use wedge[] to find
the seam pair anymore, and need to switch to using loop/loopback based
on whether the edge is forward or backward wrt original topology.

This is somewhat risky as it requires loop/loopback to stay valid and
coherent throughout the entire simplification process, whereas
previously bugs here would simply block certain collapses; now it's
theoretically possible to get an out of bounds read if the loop data for
the other wedge points nowhere. However, extensive testing and fuzzing
did not uncover any further issues with this.