simplify: Compute collapse error for each component
After computing component ids, we can compute a bounding box for each
component and derive an error value from it that will be used for
pruning.

For now we allocate an entire bounding box for each component for
performance, but we might use some other way of computing these in the
future that is less memory intensive; usually, component sizes are
fairly large so this is an acceptable use of memory.

It's not clear yet if AABB diagonal is a reasonable approximation of
component error; this is pending future testing. Note that we square
the value for consistency with quadric errors.