simplify: Switch to approximate spherical error for components
Instead of computing an AABB and using half diagonal as an error,
compute an approximate sphere bounds and use radius as an error.

This makes the error rotationally invariant (modulo inaccuracies of the
computation), and produces a more reasonable pruning order in practice
on a few test meshes, although both measurement methods have their own
issues.