simplify: Keep track of the next component error
Instead of only computing the next component error at the end of the
collapse loop, we keep track of it as we prune components. This is
fairly cheap, as usually the number of components is much smaller than
the number of indices at any point in time, and this helps us avoid
calling pruneComponents excessively: after this change, while the
behavior of simplification stays the same, only ~1% of calls to
pruneComponents end up removing no triangles (which happens when the
component was entirely removed through edge collapses).