simplify: Add a note about locked target vertices
The collapse scheduling mechanics actually permit collapsing more than
one vertex to the same target. We normally don't do this because that
may lead to collapses v0->v1 even if v1->v0 has better error, since we
haven't had a chance to recompute the error.

However, if the target vertex can not move (if it's kind is locked)
anyway, this constraint does not apply: we dutifully update quadrics of
locked vertices but these updates aren't used for any evaluation. This
could allow us to collapse more vertices in a single pass if the target
is locked.

Trying this on a set of test meshes, however, a few percentage of meshes
exhibit somewhat higher resulting error, and this error is also
apparent visually. This is probably due to an unfair advantage that
collapses to locked get; even though the error is technically correct,
we may go further in the original sorted collapse list and select edges
that we would not need to collapse as edges with smaller error become
available.

So this change is a no-op but this discovery and testing results seem
important to document for the future.