gltfpack: Deduplicate primitive accessors/data during writing
When meshes have different materials, we can not deduplicate them early
in the processing flow. Depending on the material structure, we might
end up with a different final vertex stream / index buffer, but since
the processing is deterministic often they will end up being the same.

We can thus reuse the geometry hash to cache primitive accessors and
data: if we have seen the exact same geometry before, we can reuse the
entire JSON blob (that contains accessor indices that point to buffer
views) instead of emitting new ones.

Technically, it's also possible to reuse individual streams, which is
more general, however this is more complicated to implement because
accessors with the same data may have a different setup, and is less
often useful.

For now this means that we recompute the mesh hashes redundantly again.