demo: Improve boundsError to take camera parameters into account
While boundsError is an approximation and the precise version would
compute the projected sphere bounds taking into account perspective
distortion and near plane clipping, our simpler variant can work as well
and is invariant to camera rotation which is a nice property in certain
cases; for example in VR you might want to compute LOD based on a
"third" eye, a center camera, so that the geometry is in sync between
two eyes.

Our boundsError implementation was mostly serviceable but did not depend
on camera field of view and had some potential issues with monotonicity
for cases where the camera position was inside the sphere.

We now clip the distance to znear and properly scale it so that the
result is a normalized screen space error (0..1, multiply by screen
height to get pixels). This ensures that if the bounds.error is 0,
boundsError will return 0 even if the camera is inside, which should
ensure proper cluster selection for extreme closeups.