simplify: Move vertex_error update to updateQuadrics
This way performEdgeCollapses no longer needs access to quadrics data,
and when attributes are used and we need to compute distance error, the
access to vertex quadrics is maximally local as we use them right after
aggregation, which saves 1-2% on large meshes.

For now we keep the optimization of not recomputing the distance error
unless attributes are used; it makes the data flow slightly awkward and
seems to have <1% performance impact but we can remove it later if it
becomes more problematic.