js: Update README.md
Document ErrorAbsolute and Sparse options for simplification. Sparse is
probably less useful in JS as the cost of copying the vertex data is
paid in full but would still be useful to reduce memory consumption and
improve performance; Prune is only available as an experimental option
and we currently don't document those.