clusterizer: Fix rescoring to only affect new meshlets
When getNeighborTriangle's best selection can't fit into the current
meshlet, we redo the search process using an adjusted criteria that
ignores position/normal as that is irrelevant. However, internally
getNeighborTriangle adjusted best_extra to act as a topology score
instead of just the extra count. This would lead to the last triangle of
the meshlet being reselected in addition to the first one in some cases.

Also change _extra variables to int as these are small.

Note: by itself, this change actually slightly reduces the
clusterization quality due to non-local effects of the rescoring.