gltfpack: Allow per-texture-class control for texture limit
Similarly to texture scaling, there are cases where setting individual
limits for different classes can be helpful. Using -ts works when the
input assets are known to have reasonable sizes (and allows to preserve
the size considerations for individual materials), but for a completely
arbitrary input, it can be helpful to enforce a lower limit for normal
maps or attribute maps when they use UASTC.

Since we now have a per-kind limit, the power-of-two validation is a
little more cumbersome to do; for now just remove it as it's unclear
that it's practically useful. If -tl is used to specify a
non-power-of-two limit, -tp will simply round it using the existing
logic.