gltfpack: Refactor JSON array output
A fair amount of code has to deal with manually writing arrays of
numbers to JSON; a helper function simplifies this and reduces the
potential for errors.

The only downside here is that we no longer get compile time index
bounds validation; if this ever comes up we could fix it via a template
wrapper.